### 2021-09-30 #89 @whedon
[PRE REVIEW]: Applied Measure Theory for Probabilistic Modeling
**Submitting author:** @mschauer (<a href="http://orcid.org/0000-0003-3310-7915">Moritz Schauer</a>)
**Repository:** <a href="https://github.com/cscherrer/MeasureTheory.jl" target ="_blank">https://github.com/cscherrer/MeasureTheory.jl</a>
**Version:**
