### 2021-10-01 3c4f928
fix some typos (#162)
