+++ paper/paper.tex
@@ -76,6 +76,7 @@
 %     morekeywords = [2]{Beta, Distributions}, % define more types and modules
 % }
 
+\usepackage{inconsolata}
 \usepackage[normalem]{ulem} %\sout
 \usepackage{centernot} %\centernot
 %\usepackage[colorinlistoftodos]{todonotes}
@@ -122,7 +123,8 @@ The need to preserve this computationally so important locality leads to the new
 
 \emph{Distributions are an insufficient abstraction for probabilistic modeling.}
 
-Let's first consider Bayesian modeling. In the posterior density of the parameter given the observation or data,
+Let's first consider Bayesian modeling. In the posterior density of the
+parameter $\theta$ given the observation or data $x$,
 \[
 p(\theta \mid x) = \frac{p(\theta)\ p(x \mid \theta)}{p(x)}\ ,
 \]
@@ -179,7 +181,7 @@ Some parts of our approach can been seen in existing systems, though these are s
 We'll now describe some foundations to help the reader get a deeper understanding of our approach, in particular what measures and probability distributions are and how they relate to notions such as the probability of something happening in the real world or the notion of volume of the space we are living in. 
 
 Throughout this discussion, it's important to keep in mind that for us the measure-theoretic abstractions are only means to an end.% 
-\footnote{The stackoverflow discussion \url{https://mathoverflow.net/q/11591} discusses books on measure theory, M.S.~likes \cite{Shiryaev1996} as starting point and uses \cite{Elstrodt2011} and \cite{Bauer_Heinz1982-01-01} as reference. As we anticipate interest in connecting with the actual theory, we give some pointers to more technical details in the footnotes.}
+\footnote{The stackoverflow discussion \url{https://mathoverflow.net/q/11591} discusses books on measure theory; M.S.~likes \cite{Shiryaev1996} as starting point and uses \cite{Elstrodt2011} and \cite{Bauer_Heinz1982-01-01} as reference. As we anticipate interest in connecting with the actual theory, we give some pointers to more technical details in the footnotes.}
 
 Our primary interest, and the goal of \verb|MeasureTheory|, is to offer support for \emph{applied probabilistic modeling}.
 
@@ -195,10 +197,17 @@ One further such space that deserves mention is traditionally denoted $\Omega$. 
 
 Computationally, $\Omega$ can be considered to be the set of possible states of a random number generator \verb|rng::AbstractRNG|, which is the source of computational randomness.
 
-It has a special role of tying random quantities a program produces and their distributions together, through the notion of random variables, which are functions $\bX$, $\bY$ from $\Omega$ taking values in spaces $X$, $Y$ such as $\{1,2,3,4,5,6\}$ or $\RR^3$.\footnote{Additionally, one requires that random variables are \href{https://en.wikipedia.org/wiki/Measurable_function}{\emph{measurable function}} to be one for which the inverse image of an event $\subset X$ is also an event of $\Omega$.
-
+$\Omega$ has a special role of tying random quantities a program produces and their
+distributions together, through the notion of random variables, which are
+functions $\bX$, $\bY$ from $\Omega$ taking values in spaces $X$, $Y$ such as
+$\{1,2,3,4,5,6\}$ or $\RR^3$.\footnote{Additionally, one requires a random
[…]
