### 2021-10-04 554dea3
exponential updates
