### 2021-10-04 0890e24
Fix exponential (#164)
* Make _logpdf use Real (for autodiff friendliness)

* exponential updates

* bump version
