+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.12",
   "v0.11",
   "v0.10",
   "v0.9",
@@ -12,5 +13,5 @@ var DOC_VERSIONS = [
   "v0.2",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.11.4";
+var DOCUMENTER_NEWEST = "v0.12.0";
 var DOCUMENTER_STABLE = "stable";
