### 2021-10-07 f352744
Symbolic-friendly StudentT
