### 2021-10-07 e69d17a
drop redudant code
