### 2021-10-10 743baeb
Returns{T} <: Function
