### 2021-10-11 #92 @whedon
[REVIEW]: Applied Measure Theory for Probabilistic Modeling
**Submitting author:** @mschauer (<a href="http://orcid.org/0000-0003-3310-7915">Moritz Schauer</a>)
**Repository:** <a href="https://github.com/cscherrer/MeasureTheory.jl" target ="_blank">https://github.com/cscherrer/MeasureTheory.jl</a>
**Version:** v0.16.4
