+++ src/parameterized/mvnormal.jl
@@ -5,25 +5,25 @@ export MvNormal
 
 function MvNormal(nt::NamedTuple{(:μ,)})
     dim = size(nt.μ)
-    Affine(nt, Normal() ^ dim)
+    affine(nt, Normal() ^ dim)
 end
 
 function MvNormal(nt::NamedTuple{(:σ,)})
-    dim = size(nt.σ, 2)
-    Affine(nt, Normal() ^ dim)
+    dim = colsize(nt.σ)
+    affine(nt, Normal() ^ dim)
 end
 
 function MvNormal(nt::NamedTuple{(:ω,)})
-    dim = size(nt.ω, 1)
-    Affine(nt, Normal() ^ dim)
+    dim = rowsize(nt.ω)
+    affine(nt, Normal() ^ dim)
 end
 
 function MvNormal(nt::NamedTuple{(:μ, :σ,)})
-    dim = size(nt.σ, 2)
-    Affine(nt, Normal() ^ dim)
+    dim = colsize(nt.σ)
+    affine(nt, Normal() ^ dim)
 end
 
 function MvNormal(nt::NamedTuple{(:μ, :ω,)})
-    dim = size(nt.ω, 1)
-    Affine(nt, Normal() ^ dim)
+    dim = rowsize(nt.ω)
+    affine(nt, Normal() ^ dim)
 end
