### 2021-10-18 c96fb20
Some refactoring
