+++ src/parameterized/cauchy.jl
@@ -6,14 +6,23 @@ export Cauchy, HalfCauchy
 @parameterized Cauchy(μ,σ) ≃ (1/π) * Lebesgue(ℝ)
 
 @kwstruct Cauchy()
+@kwstruct Cauchy(μ)
+@kwstruct Cauchy(σ)
+@kwstruct Cauchy(μ,σ)
+@kwstruct Cauchy(ω)
+@kwstruct Cauchy(μ,ω)
 
-Cauchy(nt::NamedTuple{(:μ,:σ)}) = affine(nt, Cauchy() ^ colsize(nt.σ))
-Cauchy(nt::NamedTuple{(:μ,:ω)}) = affine(nt, Cauchy() ^ rowsize(nt.ω))
-Cauchy(nt::NamedTuple{(:σ,)}) = affine(nt, Cauchy() ^ colsize(nt.σ))
-Cauchy(nt::NamedTuple{(:ω,)}) = affine(nt, Cauchy() ^ rowsize(nt.ω))
-Cauchy(nt::NamedTuple{(:μ,)}) = affine(nt, Cauchy() ^ size(nt.μ))
 
-@affinepars Cauchy
+
+for N in AFFINEPARS
+    @eval begin
+        proxy(d::Cauchy{$N}) = affine(params(d), Cauchy())
+        logdensity(d::Cauchy{$N}, x) = logdensity(proxy(d), x)
+        basemeasure(d::Cauchy{$N}) = basemeasure(proxy(d))
+    end
+end
+
+# @affinepars Cauchy
 
 function logdensity(d::Cauchy{()} , x) 
     return -log1p(x^2)
@@ -34,3 +43,8 @@ TV.as(::Cauchy) = asℝ
 HalfCauchy(σ) = HalfCauchy(σ=σ)
 
 distproxy(d::Cauchy{()}) = Dists.Cauchy()
+distproxy(d::Cauchy{(:μ,)}) = Dists.Cauchy(d.μ, 1.0)
+distproxy(d::Cauchy{(:σ,)}) = Dists.Cauchy(0.0, d.σ)
+distproxy(d::Cauchy{(:μ,:σ)}) = Dists.Cauchy(d.μ, d.σ)
+distproxy(d::Cauchy{(:ω,)}) = Dists.Cauchy(0.0, inv(d.ω))
+distproxy(d::Cauchy{(:μ,:ω)}) = Dists.Cauchy(d.μ, inv(d.ω))
