+++ src/parameterized/gumbel.jl
@@ -2,14 +2,25 @@
 
 export Gumbel
 
-@parameterized Gumbel() ≃ Lebesgue(ℝ)
+@parameterized Gumbel() 
+
+basemeasure(::Gumbel{()}) = Lebesgue(ℝ)
 
 @kwstruct Gumbel()
 
-Gumbel(nt::NamedTuple{(:σ,)}) = affine(nt, Gumbel())
+@kwstruct Gumbel(β)
 
-@affinepars Gumbel
+@kwstruct Gumbel(μ,β)
 
+# map affine names to those more common for Gumbel
+for N in [(:μ,), (:σ,), (:μ,:σ)]
+    G = replace(N, :σ => :β)
+    @eval begin
+        proxy(d::Gumbel{$G}) = affine(NamedTuple{$N}(values(params(d))), Gumbel())
+        logdensity(d::Gumbel{$G}, x) = logdensity(proxy(d), x)
+        basemeasure(d::Gumbel{$G}) = basemeasure(proxy(d))
+    end
+end
 
 function logdensity(d::Gumbel{()} , x)
     return -exp(-x) - x
