+++ test/runtests.jl
@@ -454,8 +454,9 @@ end
 
 @testset "Affine" begin
     testmeasures = [
-        (Cauchy, (;))
-        (Normal, (;))
+        (Normal, NamedTuple())
+        (Cauchy, NamedTuple())
+        (Laplace, NamedTuple())
         (StudentT, (ν=3,))
     ]
 
@@ -469,16 +470,20 @@ end
     for (M, nt) in testmeasures
         for p in [(μ=1,), (μ=1,σ=2), (μ=1,ω=2), (σ=2,), (ω=2,)]
             d = M(merge(nt, p))
+            @info "Testing $d"
             test_noerrors(d)
         end
-        for n in 1:3
-            for k in 1:n
-                pars = [(μ=randn(n),), (μ=randn(n),σ=randn(n,k)), (μ=randn(n),ω=randn(k,n)), (σ=randn(n,k),), (ω=randn(k,n),)]
-                for p in pars
-                    d = M(merge(nt, p))
-                    test_noerrors(d)
-                end
-            end
-        end
+        # for n in 1:3
+        #     @show n
+        #     for k in 1:n
+        #         @show k
+        #         pars = [(μ=randn(n),), (μ=randn(n),σ=randn(n,k)), (μ=randn(n),ω=randn(k,n)), (σ=randn(n,k),), (ω=randn(k,n),)]
+        #         for p in pars
+        #             @show p
+        #             d = M(merge(nt, p))
+        #             test_noerrors(d)
+        #         end
+        #     end
+        # end
     end
 end
