### 2021-10-18 7ce5d69
Actually, let's just make this `replace` stuff more direct
