### 2021-10-18 de50b0a
Update MvNormal (#160)
* update MvNormal to use Affine

* typos

* inline _logpdf

* make things easily dubuggable

* update tests

* AND ANOTHER

* update tests, but MANY NEED FIXING

* LogExpFunctions 0.3.3 (for xlog1py)

* working on tests

* logpdf from MeasureBase

* logpdf test

* Symbolic-friendly StudentT

* drop redudant code

* basekernel

* bugfix

* Returns{T} <: Function

* fixing stuff

* asparams(::Affine)

* debugging, and tests

* use Measurebase.Returns (usually from Base) instead of making another

* adjust for Productmesaure updates

* Some refactoring

* MvNormal updates


[…]
