+++ src/combinators/product.jl
@@ -1,7 +1,4 @@
-
-
-
-function TV.as(d::ProductMeasure{F,A}) where {F,A<:AbstractArray}
+function TV.as(d::ProductMeasure{F,S,A}) where {F,S,A<:AbstractArray}
     d1 = marginals(d).f(first(marginals(d).data))
     as(Array, as(d1), size(marginals(d))...)
 end
@@ -9,12 +6,14 @@ end
 ###############################################################################
 # I <: Base.Generator
 
-function TV.as(d::ProductMeasure{F,I}) where {F, I<:Base.Generator}
+function TV.as(d::ProductMeasure{F,S,I}) where {F,S,I<:Base.Generator}
     d1 = marginals(d).f(first(marginals(d).iter))
     as(Array, as(d1), size(marginals(d))...) 
 end
 
-
+function TV.as(d::ProductMeasure{Returns{T},F,A}) where {T, F, A <: AbstractArray}
+    as(Array, as(d.f.f.value), size(d.pars))
+end
 
 function Base.rand(rng::AbstractRNG, ::Type{T}, d::ProductMeasure, d1::Dists.Distribution) where {T}
     mar = marginals(d)
@@ -32,11 +31,11 @@ end
 
 
 # e.g. set(Normal(μ=2)^5, params, randn(5))
-function Accessors.set(d::ProductMeasure{F,A}, ::typeof(params), p::AbstractArray) where {F,A<:AbstractArray}
+function Accessors.set(d::ProductMeasure{F,S,A}, ::typeof(params), p::AbstractArray) where {F,S,A<:AbstractArray}
     set.(marginals(d), params, p)
 end
 
-function Accessors.set(d::ProductMeasure{F,A}, ::typeof(params), p) where {F,A<:AbstractArray}
+function Accessors.set(d::ProductMeasure{F,S,A}, ::typeof(params), p) where {F,S,A<:AbstractArray}
     par = typeof(d.pars[1])(p)
     ProductMeasure(d.f, Fill(par, size(d.pars)))
 end
