+++ src/parameterized/beta.jl
@@ -2,7 +2,7 @@
 
 export Beta
 
-@parameterized Beta(α,β) ≃ Lebesgue(𝕀)
+@parameterized Beta(α,β)
 
 @kwstruct Beta(α, β)
 
@@ -15,8 +15,20 @@ export Beta
 
 TV.as(::Beta) = as𝕀
 
-function logdensity(d::Beta{(:α, :β)}, x)
-    return xlogy(d.α - 1, x) + xlog1py(d.β - 1, -x) - logbeta(d.α, d.β)
+function logdensity(d::Beta{(:α, :β), Tuple{A,B}}, x::X) where {A,B,X}
+    if static_hasmethod(xlogy, Tuple{A,X}) && static_hasmethod(xlog1py, Tuple{B,X})
+        return xlogy(d.α - 1, x) + xlog1py(d.β - 1, -x) 
+    else
+        return (d.α - 1) * log(x) + (d.β - 1) * log1p(-x)
+    end
+end
+
+function basemeasure(d::Beta{(:α,:β)})
+    inbounds(x) = 0 < x < 1
+    constℓ = 0.0
+    varℓ() = - logbeta(d.α, d.β)
+    base = Lebesgue(ℝ)
+    FactoredBase(inbounds, constℓ, varℓ, base)
 end
 
 Base.rand(rng::AbstractRNG, T::Type, μ::Beta) = rand(rng, Dists.Beta(μ.α, μ.β))
