+++ Project.toml
@@ -21,6 +21,7 @@ MeasureBase = "fa1605e6-acd5-459c-a1e6-7e635759db14"
 NamedTupleTools = "d9ec5142-1e00-5aa0-9d6a-321866360f50"
 NestedTuples = "a734d2a7-8d68-409b-9419-626914d4061d"
 PositiveFactorizations = "85a6dd25-e78a-55b7-8502-1745935b8125"
+PrettyPrinting = "54e16d92-306c-5ea0-a30b-337be88ac337"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
