### 2021-11-02 355d417
Some refactoring for Markov chains
