+++ src/MeasureTheory.jl
@@ -43,7 +43,7 @@ import LogExpFunctions
 import NamedTupleTools
 
 import MeasureBase: testvalue, logdensity, density, basemeasure, kernel, params, ∫
-import MeasureBase: affine, supportdim
+import MeasureBase: affine, supportdim, ≪
 
 import PrettyPrinting as Pretty
 
