+++ Project.toml
@@ -47,6 +47,7 @@ MeasureBase = "0.4"
 NamedTupleTools = "0.13"
 NestedTuples = "0.3"
 PositiveFactorizations = "0.2"
+PrettyPrinting = "0.3"
 Reexport = "1"
 SpecialFunctions = "1"
 StatsFuns = "0.9"
