### 2021-11-02 10b6646
copy(rng) might change its type (e.g. GLOBAL_RNG)
