+++ Project.toml
@@ -37,13 +37,13 @@ ConstructionBase = "1.3"
 Distributions = "0.25"
 DynamicIterators = "0.4"
 FillArrays = "0.12"
-InfiniteArrays = "0.11"
+InfiniteArrays = "0.11, 0.12"
 KeywordCalls = "0.2"
 LogExpFunctions = "0.3.3"
 MLStyle = "0.4"
 MacroTools = "0.5"
 MappedArrays = "0.4"
-MeasureBase = "0.4"
+MeasureBase = "0.5"
 NamedTupleTools = "0.13"
 NestedTuples = "0.3"
 PositiveFactorizations = "0.2"
