+++ src/MeasureTheory.jl
@@ -46,7 +46,9 @@ import MeasureBase: testvalue, logdensity, density, basemeasure, kernel, params,
 import MeasureBase: affine, supportdim, ≪
 using MeasureBase: constructor
 
-import PrettyPrinting as Pretty
+import PrettyPrinting
+
+const Pretty = PrettyPrinting
 
 import Base: rand
 
