+++ src/combinators/chain.jl
@@ -1,10 +1,11 @@
 using ConcreteStructs
 using DynamicIterators
-using DynamicIterators: dub
+import DynamicIterators: dub, dyniterate, evolve
 using Base.Iterators: SizeUnknown, IsInfinite
-import DynamicIterators: dyniterate, evolve
 
 import MeasureBase: For
+
+
 
 export Chain
 
@@ -12,6 +13,10 @@ export Chain
     κ::K
     μ::M
 end
+
+Pretty.quoteof(c::Chain) = :(Chain($(Pretty.quoteof(c.κ)), $(Pretty.quoteof(c.μ))))
+
+Base.length(::Chain) = ∞
 
 function basemeasure(mc::Chain)
     Chain(basemeasure ∘ mc.κ, basemeasure(mc.μ))
@@ -39,73 +44,19 @@ dyniterate(E::Chain, ::Nothing) = dub(evolve(E))
 Base.iterate(E::Chain) = dyniterate(E, nothing)
 Base.iterate(E::Chain, value) = dyniterate(E, value)
 
-function DynamicIterators.dyniterate(r::Chain, (u,rng)::Sample)
-    μ = r.κ(u) 
-    u = rand(rng, μ)
-    return u, Sample(u, rng)
+function DynamicIterators.dyniterate(r::Chain, (x,rng)::Sample)
+    μ = r.κ(x) 
+    y = rand(rng, μ)
+    return (μ ↝ y), Sample(y, rng)
 end
+
 Base.IteratorSize(::Chain) = IsInfinite()
 Base.IteratorSize(::Type{Chain}) = IsInfinite()
 
 
-@concrete terse struct Realized{R,S,T} <: DynamicIterators.DynamicIterator
-    rng::ResettableRNG{R,S}
-    iter::T
-end
-
-Base.IteratorEltype(mc::Realized) = Base.HasEltype()
-
-function Base.eltype(::Type{Rz}) where {R,S,T,Rz <: Realized{R,S,T}}
-    eltype(T)
-end
-
-Base.length(r::Realized) = length(r.iter)
-
-Base.size(r::Realized) = size(r.iter)
-
-Base.IteratorSize(::Type{Rz}) where {R,S,T, Rz <: Realized{R,S,T}} = Base.IteratorSize(T)
-Base.IteratorSize(r::Rz) where {R,S,T, Rz <: Realized{R,S,T}} = Base.IteratorSize(r.iter)
-
-
-function Base.iterate(rv::Realized{R,S,T}) where {R,S,T}
-    if static_hasmethod(evolve, Tuple{T})
-        dyniterate(rv, nothing)
-    else
-        !isnothing(rv.rng.seed) && reset!(rv.rng)
-        μ,s = iterate(rv.iter)
-        x = rand(rv.rng, μ)
-        x,s
-    end
-end
-
-
-function Base.iterate(rv::Realized{R,S,T}, s) where {R,S,T}
-    if static_hasmethod(evolve, Tuple{T})
-        dyniterate(rv, s)
-    else
-        μs = iterate(rv.iter, s)
-        isnothing(μs) && return nothing
-        (μ,s) = μs
-        x = rand(rv.rng, μ)
-        return x,s
-    end
-end
-
-
-function dyniterate(rv::Realized, ::Nothing)
-    !isnothing(rv.rng.seed) && reset!(rv.rng)
-    μ = evolve(rv.iter)
-    x = rand(rv.rng, μ)
-    x, Sample(x, rv.rng)
-end
-function dyniterate(rv::Realized, u::Sample)
-    dyniterate(rv.iter, u)
-end
-
 function Base.rand(rng::AbstractRNG, T::Type, chain::Chain)
-    seed = rand(rng, UInt)
[…]
