### 2021-11-05 27c30cc
Include MeasureBase docs in MeasureTheory
