### 2021-11-05 7af6ad4
Set up clean documentation (#168)
* Include MeasureBase docs in MeasureTheory

* Clean up README

* Remove support in docs

* Remove useless MeasureBase import

* Version bump
