+++ docs/src/api_measurebase.md
@@ -1,0 +1,5 @@
+# MeasureBase API
+
+```@autodocs
+Modules = [MeasureBase]
+```
