### 2021-11-08 9f10340
remove likely unused method (#171)
