+++ versions.js
@@ -14,5 +14,5 @@ var DOC_VERSIONS = [
   "v0.2",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.13.1";
+var DOCUMENTER_NEWEST = "v0.13.2";
 var DOCUMENTER_STABLE = "stable";
