+++ v0.13.2/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"old_readme/#MeasureTheory","page":"MeasureTheory","title":"MeasureTheory","text":"","category":"section"},{"location":"old_readme/","page":"MeasureTheory","title":"MeasureTheory","text":"(Image: Stable) (Image: Dev) (Image: Build Status) (Image: Coverage)","category":"page"},{"location":"old_readme/","page":"MeasureTheory","title":"MeasureTheory","text":"Check out our JuliaCon submission","category":"page"},{"location":"old_readme/","page":"MeasureTheory","title":"MeasureTheory","text":"MeasureTheory.jl is a package for building and reasoning about measures.","category":"page"},{"location":"old_readme/#Why?","page":"MeasureTheory","title":"Why?","text":"","category":"section"},{"location":"old_readme/","page":"MeasureTheory","title":"MeasureTheory","text":"A distribution (as in Distributions.jl) is also called a probability measure, and carries with it the constraint of adding (or integrating) to one. Statistical work usually requires this \"at the end of the day\", but enforcing it at each step of a computation can have considerable overhead.","category":"page"},{"location":"old_readme/","page":"MeasureTheory","title":"MeasureTheory","text":"As a generalization of the concept of volume, measures also have applications outside of probability theory.","category":"page"},{"location":"old_readme/#Goals","page":"MeasureTheory","title":"Goals","text":"","category":"section"},{"location":"old_readme/#Distributions.jl-Compatibility","page":"MeasureTheory","title":"Distributions.jl Compatibility","text":"","category":"section"},{"location":"old_readme/","page":"MeasureTheory","title":"MeasureTheory","text":"Distirbutions.jl is wildly popular, and is large enough that replacing it all at once would be a major undertaking. ","category":"page"},{"location":"old_readme/","page":"MeasureTheory","title":"MeasureTheory","text":"Instead, we should aim to make any Distribution easily usable as a Measure. We'll most likely implement this using an IsMeasure trait. ","category":"page"},{"location":"old_readme/#Absolute-Continuity","page":"MeasureTheory","title":"Absolute Continuity","text":"","category":"section"},{"location":"old_readme/","page":"MeasureTheory","title":"MeasureTheory","text":"For two measures μ, ν on a set X, we say μ is absolutely continuous with respect to ν if ν(A)=0 implies μ(A)=0 for every measurable subset A of X.","category":"page"},{"location":"old_readme/","page":"MeasureTheory","title":"MeasureTheory","text":"The following are equivalent:","category":"page"},{"location":"old_readme/","page":"MeasureTheory","title":"MeasureTheory","text":"μ ≪ ν\nμ is absolutely continuous wrt ν\nThere exists a function f such that μ = ∫f dν","category":"page"},{"location":"old_readme/","page":"MeasureTheory","title":"MeasureTheory","text":"So we'll need a ≪ operator. Note that ≪ is not antisymmetric; it's common for both μ ≪ ν and  ν ≪ μ to hold. ","category":"page"},{"location":"old_readme/","page":"MeasureTheory","title":"MeasureTheory","text":"If μ ≪ ν and  ν ≪ μ, we say μ and ν are equivalent and write μ ≃ ν. (This is often written as μ ~ ν, but we reserve ~ for random variables following a distribution, as is common in the literature and probabilistic programming languages.)","category":"page"},{"location":"old_readme/","page":"MeasureTheory","title":"MeasureTheory","text":"If we collapse the equivalence classes (under ≃), ≪ becomes a partial order.","category":"page"},{"location":"old_readme/","page":"MeasureT
[…]
