### 2021-11-08 0f6d9da
Update docs/src/affine.md
Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>
