### 2021-11-08 7de094d
Update docs/src/affine.md
Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>
