### 2021-11-08 1047934
Update docs/src/affine.md
Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>
