### 2021-11-08 fd59205
Update docs/src/affine.md
Co-authored-by: Moritz Schauer <moritzschauer@web.de>
