### 2021-11-09 7102906
Doc adjustments (#169)
* Tweaks

* Further tweaks

* Update docs/src/affine.md

Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>

* Update docs/src/affine.md

* Update docs/src/affine.md

* Update docs/src/affine.md

* Update docs/src/affine.md

Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>

* Update docs/src/affine.md

* Update docs/src/affine.md

Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>

* Update docs/src/affine.md

Co-authored-by: Moritz Schauer <moritzschauer@web.de>

* Update docs/src/affine.md

Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>
Co-aut
[…]
