### 2021-11-15 760bb2e
CompatHelper: bump compat for TransformVariables to 0.5, (keep existing compat)
