+++ Project.toml
@@ -51,7 +51,7 @@ PrettyPrinting = "0.3"
 Reexport = "1"
 SpecialFunctions = "1"
 StatsFuns = "0.9"
-TransformVariables = "0.4"
+TransformVariables = "0.4, 0.5"
 Tricks = "0.1"
 julia = "1.5"
 
