+++ src/parameterized/uniform.jl
@@ -10,7 +10,7 @@ export Uniform
 ###############################################################################
 # Standard Uniform
 
-function basemeasure(::Uniform{()})
+@inline function basemeasure(::Uniform{()})
     inbounds(x) = 0 < x < 1
     constℓ = 0.0
     varℓ() = 0.0
