+++ src/parameterized/negativebinomial.jl
@@ -14,7 +14,7 @@ import Base
 ###############################################################################
 @kwstruct NegativeBinomial(r, p)
     
-function logdensity(d::NegativeBinomial{(:r, :p)}, y)
+@inline function logdensity(d::NegativeBinomial{(:r, :p)}, y)
     (r, p) = (d.r, d.p)
     return -log(y + r) - logbeta(r, y+1) + xlogy(y, p) + xlog1py(r, -p)
 end
@@ -24,7 +24,7 @@ distproxy(d::NegativeBinomial{(:r, :p)}) = Dists.NegativeBinomial(d.r, d.p)
 ###############################################################################
 @kwstruct NegativeBinomial(r, logitp)
 
-function logdensity(d::NegativeBinomial{(:r, :logitp)}, y)
+@inline function logdensity(d::NegativeBinomial{(:r, :logitp)}, y)
     (r, logitp) = (d.r, d.logitp)
     return -log(y + r) - logbeta(r, y+1) - y * log1pexp(-logitp) - r * log1pexp(logitp)
 end
@@ -36,7 +36,7 @@ distproxy(d::NegativeBinomial{(:n,:logitp)}) = Dists.NegativeBinomial(d.n, logis
 # mean λ, as in Poisson
 # Converges to Poisson as r→∞
 
-function logdensity(d::NegativeBinomial{(:r, :λ)}, y)
+@inline function logdensity(d::NegativeBinomial{(:r, :λ)}, y)
     (r, λ) = (d.r, d.λ)
     return -log(y + r) - logbeta(r, y+1) + xlogy(y, λ) + xlogx(r) - xlogy(y + r, r + λ)
 end
@@ -57,7 +57,7 @@ end
 ###############################################################################
 @kwstruct NegativeBinomial(r, logλ)
 
-function logdensity(d::NegativeBinomial{(:r, :logλ)}, y)
+@inline function logdensity(d::NegativeBinomial{(:r, :logλ)}, y)
     (r, logλ) = (d.r, d.logλ)
     λ = exp(logλ)
     return -log(y + r) - logbeta(r, y+1) + y * logλ + xlogx(r) - xlogy(y + r, r + λ)
