+++ src/parameterized/beta.jl
@@ -16,11 +16,7 @@ export Beta
 TV.as(::Beta) = as𝕀
 
 @inline function logdensity(d::Beta{(:α, :β), Tuple{A,B}}, x::X) where {A,B,X}
-    if static_hasmethod(xlogy, Tuple{A,X}) && static_hasmethod(xlog1py, Tuple{B,X})
-        return xlogy(d.α - 1, x) + xlog1py(d.β - 1, -x) 
-    else
-        return (d.α - 1) * log(x) + (d.β - 1) * log1p(-x)
-    end
+    return xlogy(d.α - 1, x) + xlog1py(d.β - 1, -x) 
 end
 
 @inline function basemeasure(d::Beta{(:α,:β)})
@@ -30,6 +26,9 @@ end
     base = Lebesgue(ℝ)
     FactoredBase(inbounds, constℓ, varℓ, base)
 end
+
+basemeasure_depth(::Beta) = static(2)
+basemeasure_depth(::Type{T}) where {T<:Beta}= static(2)
 
 Base.rand(rng::AbstractRNG, T::Type, μ::Beta) = rand(rng, Dists.Beta(μ.α, μ.β))
 
