+++ src/parameterized/laplace.jl
@@ -20,15 +20,8 @@ end
     return -abs(x)
 end
 
-Base.rand(rng::AbstractRNG, μ::Laplace{()}) = rand(rng, Dists.Laplace())
+Base.rand(rng::AbstractRNG, ::Type{T}, μ::Laplace{()}) where {T} = rand(rng, Dists.Laplace())
 
 ≪(::Laplace, ::Lebesgue{X}) where X <: Real = true
 
 TV.as(::Laplace) = asℝ
-
-distproxy(::Laplace{()}) = Dists.Laplace()
-distproxy(d::Laplace{(:μ,)}) = Dists.Laplace(d.μ, 1.0)
-distproxy(d::Laplace{(:σ,)}) = Dists.Laplace(0.0, d.σ)
-distproxy(d::Laplace{(:μ,:σ)}) = Dists.Laplace(d.μ, d.σ)
-distproxy(d::Laplace{(:ω,)}) = Dists.Laplace(0.0, inv(d.ω))
-distproxy(d::Laplace{(:μ,:ω)}) = Dists.Laplace(d.μ, inv(d.ω))
