+++ src/parameterized/negativebinomial.jl
@@ -11,6 +11,10 @@ import Base
 
 (::CountingMeasure{IntegerRange{a,b}} ≪ ::NegativeBinomial) where {a,b} = a ≥ 0 && b ≤ d.n
 
+function Base.rand(rng::AbstractRNG, ::Type{T}, d::NegativeBinomial) where {T}
+    rand(rng, proxy(d))    
+end
+
 ###############################################################################
 @kwstruct NegativeBinomial(r, p)
     
@@ -19,7 +23,7 @@ import Base
     return -log(y + r) - logbeta(r, y+1) + xlogy(y, p) + xlog1py(r, -p)
 end
 
-distproxy(d::NegativeBinomial{(:r, :p)}) = Dists.NegativeBinomial(d.r, d.p)
+proxy(d::NegativeBinomial{(:r, :p)}) = Dists.NegativeBinomial(d.r, d.p)
 
 ###############################################################################
 @kwstruct NegativeBinomial(r, logitp)
@@ -29,7 +33,7 @@ distproxy(d::NegativeBinomial{(:r, :p)}) = Dists.NegativeBinomial(d.r, d.p)
     return -log(y + r) - logbeta(r, y+1) - y * log1pexp(-logitp) - r * log1pexp(logitp)
 end
 
-distproxy(d::NegativeBinomial{(:n,:logitp)}) = Dists.NegativeBinomial(d.n, logistic(d.logitp))
+proxy(d::NegativeBinomial{(:n,:logitp)}) = Dists.NegativeBinomial(d.n, logistic(d.logitp))
 
 ###############################################################################
 @kwstruct NegativeBinomial(r, λ)
@@ -41,7 +45,7 @@ distproxy(d::NegativeBinomial{(:n,:logitp)}) = Dists.NegativeBinomial(d.n, logis
     return -log(y + r) - logbeta(r, y+1) + xlogy(y, λ) + xlogx(r) - xlogy(y + r, r + λ)
 end
 
-function distproxy(d::NegativeBinomial{(:r,:λ)})
+function proxy(d::NegativeBinomial{(:r,:λ)})
     p = d.λ / (d.r + d.λ)
     return Dists.NegativeBinomial(d.r, p)
 end
@@ -63,7 +67,7 @@ end
     return -log(y + r) - logbeta(r, y+1) + y * logλ + xlogx(r) - xlogy(y + r, r + λ)
 end
 
-function distproxy(d::NegativeBinomial{(:r,:logλ)})
+function proxy(d::NegativeBinomial{(:r,:logλ)})
     λ = exp(d.logλ)
     p = λ / (d.r + λ)
     return Dists.NegativeBinomial(d.r, p)
