+++ src/parameterized/studentt.jl
@@ -6,15 +6,16 @@ export StudentT, HalfStudentT
 @parameterized StudentT(ν)
 
 @kwstruct StudentT(ν)
-@kwstruct StudentT(ν,μ)
-@kwstruct StudentT(ν,σ)
-@kwstruct StudentT(ν,ω)
-@kwstruct StudentT(ν,μ,σ)
-@kwstruct StudentT(ν,μ,ω)
+@kwstruct StudentT(ν, μ)
+@kwstruct StudentT(ν, σ)
+@kwstruct StudentT(ν, ω)
+@kwstruct StudentT(ν, μ, σ)
+@kwstruct StudentT(ν, μ, ω)
 
 for N in AFFINEPARS
     @eval begin
-        proxy(d::StudentT{(:ν,$N...)}) = affine(NamedTupleTools.select(params(d), $N), StudentT((ν=d.ν,)))
+        proxy(d::StudentT{(:ν, $N...)}) =
+            affine(NamedTupleTools.select(params(d), $N), StudentT((ν = d.ν,)))
         logdensity(d::StudentT{(:ν, $N...)}, x) = logdensity(proxy(d), x)
         basemeasure(d::StudentT{(:ν, $N...)}) = basemeasure(proxy(d))
     end
@@ -22,7 +23,7 @@ end
 
 # @affinepars StudentT
 
-StudentT(ν, μ, σ) = StudentT((ν=ν, μ=μ, σ=σ))
+StudentT(ν, μ, σ) = StudentT((ν = ν, μ = μ, σ = σ))
 
 @kwalias StudentT [
     df       => ν
@@ -34,20 +35,20 @@ StudentT(ν, μ, σ) = StudentT((ν=ν, μ=μ, σ=σ))
     sigma    => σ
 ]
 
-@inline function logdensity(d::StudentT{(:ν,), Tuple{T}}, x::Number) where {T<:Number}
+@inline function logdensity(d::StudentT{(:ν,),Tuple{T}}, x::Number) where {T<:Number}
     ν = d.ν
-    return  xlog1py((ν + 1) / (-2), x^2 / ν)
+    return xlog1py((ν + 1) / (-2), x^2 / ν)
 end
 
-@inline function logdensity(d::StudentT{(:ν,)}, x) 
+@inline function logdensity(d::StudentT{(:ν,)}, x)
     ν = d.ν
-    return  (ν + 1) / (-2) * log1p(x^2 / ν)
+    return (ν + 1) / (-2) * log1p(x^2 / ν)
 end
 
 @inline function basemeasure(d::StudentT{(:ν,)})
     inbounds = Returns(true)
     constℓ = 0.0
-    varℓ() = loggamma((d.ν+1)/2) - loggamma(d.ν/2) - log(π * d.ν) / 2
+    varℓ() = loggamma((d.ν + 1) / 2) - loggamma(d.ν / 2) - log(π * d.ν) / 2
     base = Lebesgue(ℝ)
     FactoredBase(inbounds, constℓ, varℓ, base)
 end
@@ -57,15 +58,14 @@ TV.as(::StudentT) = asℝ
 Base.rand(rng::AbstractRNG, T::Type, μ::StudentT{(:ν,)}) = rand(rng, Dists.TDist(μ.ν))
 
 distproxy(d::StudentT{(:ν,)}) = Dists.TDist(d.ν)
-distproxy(d::StudentT{(:ν,:μ)}) = Dists.LocationScale(d.μ, 1.0, Dists.TDist(d.ν))
[…]
