### 2021-11-23 87398c2
CompatHelper: bump compat for SpecialFunctions to 2, (keep existing compat)
