+++ Project.toml
@@ -49,7 +49,7 @@ NestedTuples = "0.3"
 PositiveFactorizations = "0.2"
 PrettyPrinting = "0.3"
 Reexport = "1"
-SpecialFunctions = "1"
+SpecialFunctions = "1, 2"
 StatsFuns = "0.9"
 TransformVariables = "0.4"
 Tricks = "0.1"
