+++ src/combinators/affine.jl
@@ -1,3 +1,292 @@
+export Affine, AffineTransform
+using LinearAlgebra
+
+const AFFINEPARS = [
+    (:μ, :σ)
+    (:μ, :ω)
+    (:σ,)
+    (:ω,)
+    (:μ,)
+]
+
+struct AffineTransform{N,T}
+    par::NamedTuple{N,T}
+end
+
+quoteof(f::AffineTransform) = :(AffineTransform($(quoteof(f.par))))
+
+params(f::AffineTransform) = getfield(f, :par)
+
+@inline Base.getproperty(d::AffineTransform, s::Symbol) = getfield(getfield(d, :par), s)
+
+Base.propertynames(d::AffineTransform{N}) where {N} = N
+
+@inline Base.inv(f::AffineTransform{(:μ, :σ)}) =
+    AffineTransform((μ = -(f.σ \ f.μ), ω = f.σ))
+@inline Base.inv(f::AffineTransform{(:μ, :ω)}) = AffineTransform((μ = -f.ω * f.μ, σ = f.ω))
+@inline Base.inv(f::AffineTransform{(:σ,)}) = AffineTransform((ω = f.σ,))
+@inline Base.inv(f::AffineTransform{(:ω,)}) = AffineTransform((σ = f.ω,))
+@inline Base.inv(f::AffineTransform{(:μ,)}) = AffineTransform((μ = -f.μ,))
+
+# `size(f) == (m,n)` means `f : ℝⁿ → ℝᵐ`  
+Base.size(f::AffineTransform{(:μ, :σ)}) = size(f.σ)
+Base.size(f::AffineTransform{(:μ, :ω)}) = size(f.ω)
+Base.size(f::AffineTransform{(:σ,)}) = size(f.σ)
+Base.size(f::AffineTransform{(:ω,)}) = size(f.ω)
+
+function Base.size(f::AffineTransform{(:μ,)})
+    (n,) = size(f.μ)
+    return (n, n)
+end
+
+Base.size(f::AffineTransform, n::Int) = @inbounds size(f)[n]
+
+(f::AffineTransform{(:μ,)})(x) = x + f.μ
+(f::AffineTransform{(:σ,)})(x) = f.σ * x
+(f::AffineTransform{(:ω,)})(x) = f.ω \ x
+(f::AffineTransform{(:μ, :σ)})(x) = f.σ * x + f.μ
+(f::AffineTransform{(:μ, :ω)})(x) = f.ω \ x + f.μ
+
+rowsize(x) = ()
+rowsize(x::AbstractArray) = (size(x, 1),)
+
+function rowsize(f::AffineTransform)
+    size_f = size(f)
+
+    size_f isa Tuple{} && return 0
+    return first(size_f)
+end
+
+colsize(x) = ()
+colsize(x::AbstractArray) = (size(x, 2),)
+
+function colsize(f::AffineTransform)
+    size_f = size(f)
+
+    size_f isa NTuple{2} && return last(size_f)
+    return 0
+end
+
+@inline function apply!(x, f::AffineTransform{(:μ,)}, z)
+    x .= z .+ f.μ
+    return x
+end
+
+@inline function apply!(x, f::AffineTransform{(:σ,)}, z)
+    mul!(x, f.σ, z)
+    return x
+end
+
+@inline function apply!(x, f::AffineTransform{(:ω,),Tuple{F}}, z) where {F<:Factorization}
+    ldiv!(x, f.ω, z)
+    return x
+end
+
+@inline function apply!(x, f::AffineTransform{(:ω,)}, z)
+    ldiv!(x, factorize(f.ω), z)
+    return x
+end
+
+@inline function apply!(x, f::AffineTransform{(:μ, :σ)}, z)
+    apply!(x, AffineTransform((σ = f.σ,)), z)
[…]
