### 2021-11-30 d01db29
Val => StaticSymbol
