+++ src/parameterized/binomial.jl
@@ -7,11 +7,10 @@ using SpecialFunctions
 probit(p) = sqrt2 * erfinv(2p - 1)
 Φ(z) = (1 + erf(invsqrt2 * z)) / 2
 
-@parameterized Binomial(n, p) ≪ CountingMeasure(ℤ[0:∞])
+@parameterized Binomial(n, p) 
 
-(d::Binomial ≪ ::CountingMeasure{IntegerRange{a,b}}) where {a,b} = a ≤ 0 && b ≥ d.n
+basemeasure(d::Binomial) =  Counting(BoundedInts(static(0), d.n))
 
-(::CountingMeasure{IntegerRange{a,b}} ≪ ::Binomial) where {a,b} = a ≥ 0 && b ≤ d.n
 
 ###############################################################################
 @kwstruct Binomial(n, p)
