+++ test/runtests.jl
@@ -5,6 +5,7 @@ using Random
 using LinearAlgebra
 using DynamicIterators: trace, TimeLift
 using TransformVariables: transform, as𝕀, inverse
+using FillArrays
 
 using MeasureTheory
 
@@ -21,7 +22,7 @@ function draw2(μ)
 end
 
 function test_measure(μ)
-    logdensity(μ, testvalue(μ)) isa AbstractFloat
+    logdensity_def(μ, testvalue(μ)) isa AbstractFloat
 end
 
 test_measures = [
@@ -54,26 +55,26 @@ testbroken_measures = [
     Pushforward(as𝕀, Normal())
     # InverseGamma(2) # Not defined yet
     # MvNormal(I(3)) # Entirely broken for now
-    CountingMeasure(Float64)
+    Counting(Float64)
     Likelihood
     TrivialMeasure()
 ]
 
 @testset "testvalue" begin
     for μ in test_measures
-        # @info "testing $μ"
+        @info "testing $μ"
         @test test_measure(μ)
     end
 
     for μ in testbroken_measures
-        # @info "testing $μ"
+        @info "testing $μ"
         @test_broken test_measure(μ)
     end
     
     @testset "testvalue(::Chain)" begin
         mc =  Chain(x -> Normal(μ=x), Normal(μ=0.0))
         r = testvalue(mc)
-        @test logdensity(mc, Iterators.take(r, 10)) isa AbstractFloat
+        @test logdensity_def(mc, Iterators.take(r, 10)) isa AbstractFloat
     end
 end
 
@@ -87,11 +88,11 @@ end
         probitp = norminvcdf(p)
         y = rand(d)
 
-        ℓ = logdensity(Binomial(;n, p), y)
-        @test ℓ ≈ logdensity(Binomial(;n, logitp), y)
-        @test ℓ ≈ logdensity(Binomial(;n, probitp), y)
+        ℓ = logdensity_def(Binomial(;n, p), y)
+        @test ℓ ≈ logdensity_def(Binomial(;n, logitp), y)
+        @test ℓ ≈ logdensity_def(Binomial(;n, probitp), y)
 
-        @test_broken logdensity(Binomial(n,p), CountingMeasure(ℤ[0:n]), x) ≈ binomlogpdf(n,p,x)
+        @test_broken logdensity_def(Binomial(n,p), CountingMeasure(ℤ[0:n]), x) ≈ binomlogpdf(n,p,x)
     end
 
     @testset "Exponential" begin
@@ -101,11 +102,11 @@ end
         @test r ≈ rand(MersenneTwister(123), Exponential(logβ=log(2)))
         @test r ≈ rand(MersenneTwister(123), Exponential(logλ=log(0.5)))
 
-        ℓ = logdensity(Exponential(2), r)
-        @test ℓ ≈ logdensity(Exponential(β=2), r)
-        @test ℓ ≈ logdensity(Exponential(λ=0.5), r)
-        @test ℓ ≈ logdensity(Exponential(logβ=log(2)), r)
-        @test ℓ ≈ logdensity(Exponential(logλ=log(0.5)), r)
+        ℓ = logdensity_def(Exponential(2), r)
+        @test ℓ ≈ logdensity_def(Exponential(β=2), r)
+        @test ℓ ≈ logdensity_def(Exponential(λ=0.5), r)
+        @test ℓ ≈ logdensity_def(Exponential(logβ=log(2)), r)
+        @test ℓ ≈ logdensity_def(Exponential(logλ=log(0.5)), r)
     end
 
     @testset "NegativeBinomial" begin
[…]
