### 2021-12-01 0dd764c
more logdesnity fixes
