+++ src/combinators/transforms.jl
@@ -1,4 +1,3 @@
-using TransformVariables
 using TransformVariables: AbstractTransform, CallableTransform, CallableInverse
 
 export Pushforward
@@ -24,7 +23,7 @@ Pullback(f, ν) = Pullback(f, ν, true)
     f = pb.f
     ν = pb.ν
     if pb.logjac
-        y, logJ = transform_and_logjac(f.t, x)
+        y, logJ = TV.transform_and_logjac(f.t, x)
         return logdensity_def(ν, y) + logJ
     else
         y = f(x)
@@ -35,32 +34,32 @@ end
 @inline function logdensity_def(pf::Pushforward{F}, y) where {F<:CallableTransform}
     f = pf.f
     μ = pf.μ
-    x = inverse(f.t)(y)
+    x = TV.inverse(f.t)(y)
     if pf.logjac
-        _, logJ = transform_and_logjac(f.t, x)
+        _, logJ = TV.transform_and_logjac(f.t, x)
         return logdensity_def(μ, x) - logJ
     else
         return logdensity_def(μ, x)
     end
 end
 
-Pullback(f::AbstractTransform, ν, logjac::Bool = true) = Pullback(transform(f), ν, logjac)
+Pullback(f::AbstractTransform, ν, logjac::Bool = true) = Pullback(TV.transform(f), ν, logjac)
 Pushforward(f::AbstractTransform, ν, logjac::Bool = true) =
-    Pushforward(transform(f), ν, logjac)
+    Pushforward(TV.transform(f), ν, logjac)
 
 Pullback(f::CallableInverse, ν, logjac::Bool = true) =
-    Pushforward(transform(f.t), ν, logjac)
+    Pushforward(TV.transform(f.t), ν, logjac)
 
 Pushforward(f::CallableInverse, ν, logjac::Bool = true) =
-    Pullback(transform(f.t), ν, logjac)
+    Pullback(TV.transform(f.t), ν, logjac)
 
 Base.rand(rng::AbstractRNG, T::Type, ν::Pushforward) = ν.f(rand(rng, T, ν.μ))
 
 Base.rand(rng::AbstractRNG, T::Type, μ::Pullback) = μ.f(rand(rng, T, μ.ν))
 
-testvalue(ν::Pushforward) = transform(ν.f, testvalue(ν.μ))
+testvalue(ν::Pushforward) = TV.transform(ν.f, testvalue(ν.μ))
 
-testvalue(μ::Pullback) = transform(inverse(μ.f), testvalue(μ.ν))
+testvalue(μ::Pullback) = TV.transform(TV.inverse(μ.f), testvalue(μ.ν))
 
 basemeasure(μ::Pullback) = Pullback(μ.f, basemeasure(μ.ν), false)
 
@@ -68,7 +67,7 @@ basemeasure(ν::Pushforward) = Pushforward(ν.f, basemeasure(ν.μ), false)
 
 TV.as(ν::Pushforward) = ν.f ∘ as(ν.μ)
 
-TV.as(μ::Pullback) = inverse(μ.f) ∘ μ.ν
+TV.as(μ::Pullback) = TV.inverse(μ.f) ∘ μ.ν
 
 TV.as(::Lebesgue) = asℝ
 
