+++ src/MeasureTheory.jl
@@ -42,12 +42,14 @@ using SpecialFunctions
 import LogExpFunctions
 import NamedTupleTools
 
-import MeasureBase: testvalue, logdensity_def, density_def, basemeasure, kernel, params, paramnames, ∫
+import MeasureBase:
+    testvalue, logdensity_def, density_def, basemeasure, kernel, params, paramnames, ∫
 import MeasureBase: ≪
 using MeasureBase: constructor
 using MeasureBase: MapsTo, ↦
 using MeasureBase: BoundedInts, BoundedReals, CountingMeasure, IntegerDomain, IntegerNumbers
 using MeasureBase: weightedmeasure, restrict
+using MeasureBase: AbstractKernel
 
 import PrettyPrinting
 
