+++ src/combinators/chain.jl
@@ -116,9 +116,13 @@ function dyniterate(df::DynamicFor, st, args...)
     return (df.κ(val), state)
 end
 
-For(f, it::DynamicIterator) = DynamicFor(f, it)
+For(f, it::DynamicIterator) = DynamicFor(kernel(f), it)
 
-For(f, it::DynamicFor) = DynamicFor(f, it)
+For(f, it::DynamicFor) = DynamicFor(kernel(f), it)
+
+For(f::AbstractKernel, it::DynamicIterator) = DynamicFor(f, it)
+
+For(f::AbstractKernel, it::DynamicFor) = DynamicFor(f, it)
 
 function dyniterate(df::DynamicFor, state)
     ϕ = dyniterate(df.iter, state)
