+++ src/combinators/affine.jl
@@ -179,90 +179,86 @@ Base.size(d::Affine) = size(d.μ)
 Base.size(d::Affine{(:σ,)}) = (size(d.σ, 1),)
 Base.size(d::Affine{(:ω,)}) = (size(d.ω, 2),)
 
-logdensity_def(d::Affine, x::MapsTo) = logdensity_def(d.parent, x.x)
-
 @inline function logdensity_def(d::Affine{(:σ,)}, x)
     z = d.σ \ x
-    # @show z
-    # println()
-    logdensity_def(d.parent, z)
+    logdensityof(d.parent, z)
 end
 
 @inline function logdensity_def(d::Affine{(:ω,)}, x)
     z = d.ω * x
-    logdensity_def(d.parent, z)
+    logdensityof(d.parent, z)
 end
 
 @inline function logdensity_def(d::Affine{(:μ,)}, x)
     z = x - d.μ
-    logdensity_def(d.parent, z)
+    logdensityof(d.parent, z)
 end
 
 @inline function logdensity_def(d::Affine{(:μ, :σ)}, x)
     z = d.σ \ (x - d.μ)
-    logdensity_def(d.parent, z)
+    logdensityof(d.parent, z)
 end
 
 @inline function logdensity_def(d::Affine{(:μ, :ω)}, x)
     z = d.ω * (x - d.μ)
-    logdensity_def(d.parent, z)
+    logdensityof(d.parent, z)
 end
 
-@inline function logdensity_tuple(d::Affine{(:σ,)}, x)
-    z = d.σ \ x
-    # @show z
-    # println()
-    (logdensity_def(d.parent, z), basemeasure(d), z ↦ x)
-end
+# @inline function logdensity_tuple(d::Affine{(:σ,)}, x)
+#     z = d.σ \ x
+#     # @show z
+#     # println()
+#     (logdensity_def(d.parent, z), basemeasure(d), z ↦ x)
+# end
 
-@inline function logdensity_tuple(d::Affine{(:ω,)}, x)
-    z = d.ω * x
-    (logdensity_def(d.parent, z), basemeasure(d), z ↦ x)
-end
+# @inline function logdensity_tuple(d::Affine{(:ω,)}, x)
+#     z = d.ω * x
+#     (logdensity_def(d.parent, z), basemeasure(d), z ↦ x)
+# end
 
-@inline function logdensity_tuple(d::Affine{(:μ,)}, x)
-    z = x - d.μ
-    (logdensity_def(d.parent, z), basemeasure(d), z ↦ x)
-end
-
-@inline function logdensity_tuple(d::Affine{(:μ, :σ)}, x)
-    z = d.σ \ (x - d.μ)
-    (logdensity_def(d.parent, z), basemeasure(d), z ↦ x)
-end
-
-@inline function logdensity_tuple(d::Affine{(:μ, :ω)}, x)
-    z = d.ω * (x - d.μ)
-    (logdensity_def(d.parent, z), basemeasure(d), z ↦ x)
-end
-
-for p in AFFINEPARS
-    @eval begin
-        @inline function logdensity_tuple(d::Affine{$p}, (z, x)::MapsTo)
-            (logdensity_def(d.parent, z), basemeasure(d), z ↦ x)
-        end
-    end
-end
-
-# # logdensity_def(d::Affine{(:μ,:ω)}, x) = logdensity_def(d.parent, d.σ \ (x - d.μ))
-# @inline function logdensity_def(d::Affine{(:μ,:σ), P, Tuple{V,M}}, x) where {P, V<:AbstractVector, M<:AbstractMatrix}
+# @inline function logdensity_tuple(d::Affine{(:μ,)}, x)
 #     z = x - d.μ
[…]
