+++ src/combinators/product.jl
@@ -1,4 +1,4 @@
-function TV.as(d::ProductMeasure{F,S,A}) where {F,S,A<:AbstractArray}
+function TV.as(d::ProductMeasure{A}) where {A<:AbstractArray}
     d1 = marginals(d).f(first(marginals(d).data))
     as(Array, as(d1), size(marginals(d))...)
 end
@@ -6,21 +6,17 @@ end
 ###############################################################################
 # I <: Base.Generator
 
-function TV.as(d::ProductMeasure{K,I}) where {K,I<:Base.Generator}
+function TV.as(d::ProductMeasure{<:Base.Generator})
     d1 = marginals(d).f(first(marginals(d).iter))
     as(Array, as(d1), size(marginals(d))...)
 end
 
-function TV.as(d::ProductMeasure{Returns{T},F,A}) where {T,F,A<:AbstractArray}
-    as(Array, as(d.f.f.value), size(d.xs))
-end
-
 function Base.rand(
     rng::AbstractRNG,
-    ::Type{T},
-    d::ProductMeasure,
+    ::Type,
+    d::ProductMeasure{A},
     d1::Dists.Distribution,
-) where {T}
+) where {A<:AbstractArray}
     mar = marginals(d)
 
     # Distributions doens't (yet) have the three-argument form
@@ -28,7 +24,7 @@ function Base.rand(
 
     sz = size(mar)
     x = Array{elT,length(sz)}(undef, sz)
-    for (j, parj) in enumerate(d.xs)
+    for (j, parj) in enumerate(mar)
         x[j] = rand(rng, d.f(parj))
     end
     x
@@ -36,18 +32,18 @@ end
 
 # e.g. set(Normal(μ=2)^5, params, randn(5))
 function Accessors.set(
-    d::ProductMeasure{F,S,A},
+    d::ProductMeasure{A},
     ::typeof(params),
     p::AbstractArray,
-) where {F,S,A<:AbstractArray}
+) where {A<:AbstractArray}
     set.(marginals(d), params, p)
 end
 
 function Accessors.set(
-    d::ProductMeasure{F,S,A},
+    d::ProductMeasure{A},
     ::typeof(params),
     p,
-) where {F,S,A<:AbstractArray}
+) where {M,A<:AbstractArray{M}}
     par = typeof(d.xs[1])(p)
     ProductMeasure(d.f, Fill(par, size(d.xs)))
 end
