+++ Project.toml
@@ -5,6 +5,7 @@ version = "0.13.2"
 
 [deps]
 Accessors = "7d9f7c33-5ae7-4f3b-8dc6-eff91059b697"
+Compat = "34da2185-b29b-5c13-b0c7-acf172513d20"
 ConstructionBase = "187b0558-2788-49d3-abe0-74a17ed4e7c9"
 DensityInterface = "b429d917-457f-4dbc-8f4c-0cc954292b1d"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
@@ -33,6 +34,7 @@ Tricks = "410a4b4d-49e4-4fbc-ab6d-cb71b17b3775"
 
 [compat]
 Accessors = "0.1"
+Compat = "3.35"
 ConstructionBase = "1.3"
 DensityInterface = "0.4"
 Distributions = "0.25"
