### 2021-12-10 cscherrer/MeasureTheory.jl#179 @ludgerpaehler
Statement of Need
Per Review Criteria

> Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I believe what would really help here is to more clearly establish the rules underlying the definition of a measure in the "Why", hence leading into the definition of a probability measure essentially adding the probability constraint. As is, the "Why" is a little hard to read for someone who does not already know measure theory.
