### 2021-12-10 cscherrer/MeasureTheory.jl#180 @ludgerpaehler
Example Usage
Per the review criteria:

> Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

As is right now, the documentation is really focused on the base layer and the ability to define a new measure, but the documentation is missing a more extensive (practical) use-case. In the "Why" of the package its necessity in "Bayesian modeling" is pointed out, such example could really enhance the documentation, and make the package much more accessible to new users.
