### 2021-12-10 #92 @ludgerpaehler
Really interesting package with a strong necessity for it.

Working through the checklist, there are two points which jumped my eyes:

> Statement of Need: Do the authors clearly state what problems the software is designed to solve (..)?

https://github.com/cscherrer/MeasureTheory.jl/issues/179

> Example Usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

https://github.com/cscherrer/MeasureTheory.jl/issues/180
