### 2021-12-10 #92 @ludgerpaehler
I would furthermore encourage the authors to improve the reading flow of the paper. Two sections that the eye in this respect are

In "Why Measures?"
> Also, there’s an elegant correspondence between frequentist
and Bayesian methods, where regularization corresponds to a prior.

It is a really good analogy, but reads disconnected from the rest of the paragraph. Integration with the rest of the paragraph would elevate the reading flow greatly.

- "Related Work" just lists the 5 related papers in their own respective paragraphs, without any leading thread between the 5 of them.
