### 2021-12-10 0e74fa3
drop static weight for now
