### 2021-12-10 b7f7f84
fix sampling from ProductMeasure{<:Base.Generator}
