+++ src/realized.jl
@@ -157,3 +157,7 @@ Base.length(r::RealizedSamples) = length(r.parent)
 # Base.size(r::RealizedSamples) = ...
 
 Base.IteratorSize(r::RealizedSamples) = Base.IteratorSize(r.parent)
+
+function Base.rand(rng::AbstractRNG, ::Type{T}, d::ProductMeasure{G}) where {T, G<:Base.Generator}
+    RealizedSamples(ResettableRNG(rng), marginals(d))
+end
