+++ test/runtests.jl
@@ -385,7 +385,7 @@ end
         d = For((j^2 for j in 1:10)) do i Poisson(i) end
         x = Vector{Int16}(undef, 10)
         @test rand!(d,x) isa Vector
-        @test rand(d) isa Base.Generator
+        @test rand(d) isa MeasureTheory.RealizedSamples
     end
 
     @testset "Indexed by multiple Ints" begin
