+++ src/combinators/affine.jl
@@ -128,9 +128,11 @@ struct OrthoLebesgue{N,T} <: AbstractMeasure
     OrthoLebesgue(nt::NamedTuple{N,T}) where {N,T} = new{N,T}(nt)
 end
 
+
+
 basemeasure(d::OrthoLebesgue) = d
 
-basemeasure_depth(::Type{O}) where {O<:OrthoLebesgue} = static(0)
+tbasemeasure_depth(::Type{O}) where {O<:OrthoLebesgue} = static(0)
 
 logdensity_def(::OrthoLebesgue, x) = static(0)
 
@@ -147,6 +149,10 @@ function testvalue(d::Affine)
     f = getfield(d, :f)
     z = testvalue(parent(d))
     return f(z)
+end
+
+function tbasemeasure_type(::Type{Affine{N,M,T}}) where {N,M,T}
+    OrthoLebesgue{N,T}
 end
 
 Affine(nt::NamedTuple, μ::AbstractMeasure) = affine(nt, μ)
