### 2021-12-14 #92 @femtomc
@cscherrer
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Do you mind adopting a specific set of guidelines for contributions, and placing them in the repository? (e.g. ColPrac, or something else).
