+++ src/MeasureTheory.jl
@@ -46,7 +46,6 @@ import MeasureBase:
     testvalue, logdensity_def, density_def, basemeasure, kleisli, params, paramnames, ∫, 𝒹, ∫exp
 import MeasureBase: ≪, basemeasure_type
 import MeasureBase: tbasemeasure_depth, tbasemeasure_type
-using MeasureBase: constructor
 using MeasureBase: BoundedInts, BoundedReals, CountingMeasure, IntegerDomain, IntegerNumbers
 using MeasureBase: weightedmeasure, restrict
 using MeasureBase: AbstractKleisli
