+++ src/parameterized/laplace.jl
@@ -18,10 +18,10 @@ end
 logdensity_def(d::Laplace, x) = logdensity_def(proxy(d), x)
 
 
-basemeasure(::Laplace{()}) = (1 / 2) * Lebesgue(ℝ)
+basemeasure(::Laplace{()}) = WeightedMeasure(static(-logtwo), Lebesgue(ℝ))
 basemeasure(d::Laplace) = basemeasure(proxy(d))
 
-function basemeasure_type(::Type{<:Laplace{()}}) 
+function tbasemeasure_type(::Type{<:Laplace{()}}) 
     WeightedMeasure{Float64, Lebesgue{MeasureBase.RealNumbers}}
 end
 
