+++ src/parameterized/cauchy.jl
@@ -12,14 +12,14 @@ export Cauchy, HalfCauchy
 @kwstruct Cauchy(ω)
 @kwstruct Cauchy(μ, ω)
 
-proxy(d::Cauchy) = affine(params(d), Cauchy())
 
 logdensity_def(d::Cauchy, x) = logdensity_def(proxy(d), x)
 
-basemeasure(d::Cauchy) = (1 / π) * Lebesgue(ℝ)
+basemeasure(d::Cauchy) = WeightedMeasure(static(-logπ), Lebesgue(ℝ))
 
-basemeasure_depth(::Cauchy) = static(2)
-basemeasure_depth(::Type{C}) where {C<:Cauchy} = static(2)
+function tbasemeasure_type(::Type{Cauchy{(), Tuple{}}}) 
+    WeightedMeasure{StaticFloat64{-logπ}, Lebesgue{MeasureBase.RealNumbers}}
+end
 
 # @affinepars Cauchy
 
@@ -27,7 +27,7 @@ basemeasure_depth(::Type{C}) where {C<:Cauchy} = static(2)
     return -log1p(x^2)
 end
 
-function density(d::Cauchy{()}, x)
+function density_def(d::Cauchy{()}, x)
     return inv(1 + x^2)
 end
 
@@ -35,9 +35,12 @@ Base.rand(rng::AbstractRNG, T::Type, μ::Cauchy{()}) = randn(rng, T) / randn(rng
 
 for N in AFFINEPARS
     @eval begin
+        proxy(d::Cauchy{$N}) = affine(params(d), Cauchy())
+        @useproxy Cauchy{$N}
         rand(rng::AbstractRNG, ::Type{T}, d::Cauchy{$N}) where {T} = rand(rng, T, affine(params(d), Cauchy()))
     end
 end
+
 
 ≪(::Cauchy, ::Lebesgue{X}) where {X<:Real} = true
 
@@ -48,8 +51,3 @@ TV.as(::Cauchy) = asℝ
 HalfCauchy(σ) = HalfCauchy(σ = σ)
 
 distproxy(d::Cauchy{()}) = Dists.Cauchy()
-distproxy(d::Cauchy{(:μ,)}) = Dists.Cauchy(d.μ, 1.0)
-distproxy(d::Cauchy{(:σ,)}) = Dists.Cauchy(0.0, d.σ)
-distproxy(d::Cauchy{(:μ, :σ)}) = Dists.Cauchy(d.μ, d.σ)
-distproxy(d::Cauchy{(:ω,)}) = Dists.Cauchy(0.0, inv(d.ω))
-distproxy(d::Cauchy{(:μ, :ω)}) = Dists.Cauchy(d.μ, inv(d.ω))
