+++ src/parameterized/beta.jl
@@ -19,16 +19,18 @@ TV.as(::Beta) = as𝕀
     return xlogy(d.α - 1, x) + xlog1py(d.β - 1, -x)
 end
 
+function tbasemeasure_type(::Type{Beta{(:α, :β), T}}) where {T}
+    R = float(eltype(T))
+    FactoredBase{Base.Fix2{typeof(in), MeasureBase.BoundedReals{Static.StaticFloat64{0.0}, Static.StaticFloat64{1.0}}}, R, Returns{R}, Lebesgue{MeasureBase.RealNumbers}}
+end
+
 @inline function basemeasure(d::Beta{(:α, :β)})
-    inbounds(x) = 0 < x < 1
+    inbounds = in(𝕀)
     constℓ = 0.0
-    varℓ() = -logbeta(d.α, d.β)
+    varℓ = Returns(-logbeta(d.α, d.β))
     base = Lebesgue(ℝ)
     FactoredBase(inbounds, constℓ, varℓ, base)
 end
-
-basemeasure_depth(::Beta) = static(2)
-basemeasure_depth(::Type{T}) where {T<:Beta} = static(2)
 
 Base.rand(rng::AbstractRNG, T::Type, μ::Beta) = rand(rng, Dists.Beta(μ.α, μ.β))
 
