+++ src/parameterized/normal.jl
@@ -36,8 +36,6 @@ for N in AFFINEPARS
     end
 end
 
-
-
 @inline logdensity_def(d::Normal{()}, x) = -x^2 / 2
 @inline basemeasure(::Normal{()}) = WeightedMeasure(static(-0.5 * log2π), Lebesgue(ℝ))
 
