+++ src/parameterized/uniform.jl
@@ -9,12 +9,18 @@ export Uniform
 ###############################################################################
 # Standard Uniform
 
+
 @inline function basemeasure(::Uniform{()})
-    inbounds(x) = 0 < x < 1
-    constℓ = 0.0
-    varℓ() = 0.0
+    inbounds = in(𝕀)
+    constℓ = static(0.0)
+    varℓ = Returns(static(0.0))
     base = Lebesgue(ℝ)
     FactoredBase(inbounds, constℓ, varℓ, base)
+end
+
+
+function tbasemeasure_type(::Type{Uniform{(), Tuple{}}}) 
+    FactoredBase{typeof(in(𝕀)), Static.StaticFloat64{0.0}, Returns{Static.StaticFloat64{0.0}}, Lebesgue{MeasureBase.RealNumbers}}
 end
 
 distproxy(::Uniform{()}) = Dists.Uniform()
