### 2021-12-24 974eff4
fix Dirichlet
