+++ test/runtests.jl
@@ -407,7 +407,7 @@ end
 @testset "Density measures and Radon-Nikodym" begin
     x = randn()
     let d = ∫(𝒹(Cauchy(), Normal()), Normal())
-        @test logdensityof(d, Cauchy(), x) ≈ 0 atol=1e-12
+        @test logdensityof(𝒹(d, Cauchy()), x) ≈ 0 atol=1e-12
     end
 
     let f = 𝒹(∫(x -> x^2, Normal()), Normal())
@@ -428,21 +428,21 @@ end
         d = Normal(σ=3)
         h = HalfNormal(3)
         x = rand(h)
-        @test densityof(h, Lebesgue(ℝ), x) ≈ 2 * densityof(d, Lebesgue(ℝ), x)
+        @test densityof(𝒹(h, Lebesgue(ℝ)), x) ≈ 2 * densityof(𝒹(d, Lebesgue(ℝ)), x)
     end
 
     @testset "HalfCauchy" begin
         d = Cauchy(σ=3)
         h = HalfCauchy(3)
         x = rand(h)
-        @test densityof(h, Lebesgue(ℝ), x) ≈ 2 * densityof(d, Lebesgue(ℝ), x)
+        @test densityof(𝒹(h, Lebesgue(ℝ)), x) ≈ 2 * densityof(𝒹(d, Lebesgue(ℝ)), x)
     end
 
     @testset "HalfStudentT" begin
         d = StudentT(ν=2, σ=3)
         h = HalfStudentT(2, 3)
         x = rand(h)
-        @test densityof(h, Lebesgue(ℝ), x) ≈ 2 * densityof(d, Lebesgue(ℝ), x)
+        @test densityof(𝒹(h, Lebesgue(ℝ)), x) ≈ 2 * densityof(𝒹(d, Lebesgue(ℝ)), x)
     end
 end
 
