### 2021-12-28 c4b546f
tighten down inference
