### 2021-12-29 e1050f6
as(::PowerMeasure)
