+++ src/combinators/product.jl
@@ -1,3 +1,7 @@
+function TV.as(d::PowerMeasure)
+    as(Array, as(d.parent), length.(d.axes)...)
+end
+
 function TV.as(d::ProductMeasure{A}) where {A<:AbstractArray}
     d1 = marginals(d).f(first(marginals(d).data))
     as(Array, as(d1), size(marginals(d))...)
