+++ src/MeasureTheory.jl
@@ -41,10 +41,12 @@ using SpecialFunctions
 
 import LogExpFunctions
 import NamedTupleTools
+import InverseFunctions: inverse
+export inverse
 
 import MeasureBase:
     testvalue, logdensity_def, density_def, basemeasure, kleisli, params, paramnames, ∫, 𝒹, ∫exp
-import MeasureBase: ≪, basemeasure_type
+import MeasureBase: ≪
 using MeasureBase: BoundedInts, BoundedReals, CountingMeasure, IntegerDomain, IntegerNumbers
 using MeasureBase: weightedmeasure, restrict
 using MeasureBase: AbstractKleisli
