### 2022-01-03 f860af6
drop old tbasemeasure_type code
