+++ src/parameterized/studentt.jl
@@ -26,10 +26,6 @@ Base.rand(rng::AbstractRNG, ::Type{T}, μ::StudentT) where {T} = rand(rng, T, pr
 
 # @affinepars StudentT
 
-function tbasemeasure_type(::Type{StudentT{(:ν,), T}}) where {T}
-    FactoredBase{Returns{Bool}, Float64, MeasureTheory.var"#varℓ#10"{StudentT{(:ν,), T}}, Lebesgue{MeasureBase.RealNumbers}}
-end
-
 StudentT(ν, μ, σ) = StudentT((ν = ν, μ = μ, σ = σ))
 
 @kwalias StudentT [
