+++ src/parameterized/normal.jl
@@ -29,19 +29,11 @@ for N in AFFINEPARS
     @eval begin
         proxy(d::Normal{$N}) = affine(params(d), Normal())
         @useproxy Normal{$N}
-
-        function tbasemeasure_type(::Type{Normal{$N,T}}) where {T} 
-            Affine{$N, MeasureBase.LebesgueMeasure, T}
-        end
     end
 end
 
-@inline logdensity_def(d::Normal{()}, x) = -x^2 / static(2.0)
+@inline logdensity_def(d::Normal{()}, x) = -x^2 / 2
 @inline basemeasure(::Normal{()}) = WeightedMeasure(static(-0.5 * log2π), Lebesgue(ℝ))
-
-function tbasemeasure_type(::Type{Normal{(), Tuple{}}}) 
-    WeightedMeasure{StaticFloat64{-0.5 * log2π}, Lebesgue{MeasureBase.RealNumbers}}
-end
 
 @kwstruct Normal(μ)
 @kwstruct Normal(σ)
