+++ src/combinators/affine.jl
@@ -131,8 +131,6 @@ end
 
 basemeasure(d::OrthoLebesgue) = d
 
-tbasemeasure_depth(::Type{O}) where {O<:OrthoLebesgue} = static(0)
-
 logdensity_def(::OrthoLebesgue, x) = static(0)
 
 struct Affine{N,M,T} <: AbstractMeasure
@@ -148,10 +146,6 @@ function testvalue(d::Affine)
     f = getfield(d, :f)
     z = testvalue(parent(d))
     return f(z)
-end
-
-function tbasemeasure_type(::Type{Affine{N,M,T}}) where {N,M,T}
-    OrthoLebesgue{N,T}
 end
 
 Affine(nt::NamedTuple, μ::AbstractMeasure) = affine(nt, μ)
