### 2022-01-03 edcabfb
update logdensity_def(d::Affine{(:σ,)}, x)
