### 2022-01-06 34eab67
Move `solve` to MeasureTheory
