### 2022-01-06 cabd15b
better type stability
