+++ src/combinators/chain.jl
@@ -78,8 +78,8 @@ function Base.rand(rng::AbstractRNG, T::Type, df::DynamicFor)
     return RealizedSamples(r, df)
 end
 
-@inline function logdensity_def(df::DynamicFor, y)
-    ℓ = 0.0
+@inline function logdensity_def(df::DynamicFor{M}, y::T) where {M,T}
+    ℓ = zero(float(Core.Compiler.return_type(logdensity_def, Tuple{M,T})))
     for (xj, yj) in zip(df.iter, y)
         ℓ += logdensity_def(df.κ(xj), yj)
     end
