### 2022-01-07 abce501
work on exponential families
