+++ src/combinators/exponential-families.jl
@@ -1,32 +1,45 @@
-struct ExponentialFamily{M,H,T,A} <: AbstractKleisli
-    parent::M
+export ExponentialFamily
+
+struct ExponentialFamily{B,Θ,H,T,A} <: AbstractKleisli
+    base::B
+    θ::Θ
     η::H
     t::T
     a::A
 end
 
-function (fam::ExponentialFamily)(θ)
+
+function MeasureBase.powermeasure(fam::ExponentialFamily, dims::NTuple{N,I}) where {N,I}
+    @inline t(x) = sum(fam.t, x)
+    a = AffineTransform((σ=prod(dims),)) ∘ fam.a
+    ExponentialFamily(fam.base ^ dims, fam.θ, fam.η, t, a)
+end
+
+function (fam::ExponentialFamily)(par)
+    θ = fam.θ(par)
     η = fam.η(θ)
     a = fam.a(θ)
-    ℓ(x) = dot(η, fam.t(x)) - a
-    ∫exp(ℓ, fam.parent)
+    @inline ℓ(x) = mydot(η, fam.t(x)) - a
+    ∫exp(ℓ, fam.base)
 end
 
-struct ExpFamLikelihood{C,H,T,A} <: AbstractLikelihood
+struct ExpFamLikelihood{C,Θ,H,T,A} <: AbstractLikelihood
     c::C
+    θ::Θ
     η::H
     t::T
     a::A
 end
 
+export likelihood
 
 function likelihood(fam::ExponentialFamily, x)
-    c = logdensity_def(fam.parent, x)
+    c = logdensityof(fam.base, x)
     t = fam.t(x)
-    ExpFamLikelihood(c, fam.η, t, fam.a)
+    ExpFamLikelihood(c, fam.θ, fam.η, t, fam.a)
 end
 
-function logdensity_def(ℓ::ExpFamLikelihood, θ)
-    dot(ℓ.η(θ), ℓ.t) - ℓ.a(θ) + ℓ.c 
+@inline function logdensity_def(ℓ::ExpFamLikelihood, par)
+    θ = ℓ.θ(par)
+    mydot(ℓ.η(θ), ℓ.t) - ℓ.a(θ) + ℓ.c 
 end
-
