### 2022-01-12 35ae6c8
applying an exponential family
