+++ Project.toml
@@ -47,7 +47,7 @@ MeasureBase = "0.5"
 NamedTupleTools = "0.13"
 NestedTuples = "0.3"
 PositiveFactorizations = "0.2"
-PrettyPrinting = "0.3"
+PrettyPrinting = "0.3, 0.4"
 Reexport = "1"
 SpecialFunctions = "1"
 StatsFuns = "0.9"
