### 2022-01-20 95c443d
kwstruct for Bernoulli
