### 2022-01-20 b0c897f
Simplify logdensity_def(::DynamicFor{M}, y) for now
