+++ src/combinators/chain.jl
@@ -79,7 +79,8 @@ function Base.rand(rng::AbstractRNG, T::Type, df::DynamicFor)
 end
 
 @inline function logdensity_def(df::DynamicFor{M}, y::T) where {M,T}
-    ℓ = zero(float(Core.Compiler.return_type(logdensity_def, Tuple{M,T})))
+    # ℓ = zero(float(Core.Compiler.return_type(logdensity_def, Tuple{M,T})))
+    ℓ = 0.0
     for (xj, yj) in zip(df.iter, y)
         ℓ += logdensity_def(df.κ(xj), yj)
     end
