+++ Project.toml
@@ -7,6 +7,7 @@ version = "0.13.2"
 Accessors = "7d9f7c33-5ae7-4f3b-8dc6-eff91059b697"
 ChangesOfVariables = "9e997f8a-9a97-42d5-a9f1-ce6bfc15e2c0"
 Compat = "34da2185-b29b-5c13-b0c7-acf172513d20"
+ConcreteStructs = "2569d6c7-a4a2-43d3-a901-331e8e4be471"
 ConstructionBase = "187b0558-2788-49d3-abe0-74a17ed4e7c9"
 DensityInterface = "b429d917-457f-4dbc-8f4c-0cc954292b1d"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
@@ -16,6 +17,7 @@ Infinities = "e1ba4f0e-776d-440f-acd9-e1d2e9742647"
 InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
 InverseFunctions = "3587e190-3f89-42d0-90ee-14403ec27112"
 KeywordCalls = "4d827475-d3e4-43d6-abe3-9688362ede9f"
+LazyArrays = "5078a376-72f3-5289-bfd5-ec5146d43c02"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 LogExpFunctions = "2ab3a3ac-af41-5b50-aa03-7779005ae688"
 MLStyle = "d8e11817-5142-5d16-987a-aa16d5891078"
@@ -39,6 +41,7 @@ Tricks = "410a4b4d-49e4-4fbc-ab6d-cb71b17b3775"
 Accessors = "0.1"
 ChangesOfVariables = "0.1"
 Compat = "3.35"
+ConcreteStructs = "0.2"
 ConstructionBase = "1.3"
 DensityInterface = "0.4"
 Distributions = "0.25"
@@ -47,6 +50,7 @@ FillArrays = "0.12"
 Infinities = "0.1"
 InverseFunctions = "0.1"
 KeywordCalls = "0.2"
+LazyArrays = "0.22"
 LogExpFunctions = "0.3.3"
 MLStyle = "0.4"
 MacroTools = "0.5"
@@ -59,6 +63,7 @@ PrettyPrinting = "0.3"
 Reexport = "1"
 SpecialFunctions = "1"
 Static = "0.4"
+StaticArrays = "1.3"
 StatsFuns = "0.9"
 TransformVariables = "0.4"
 Tricks = "0.1"
