+++ src/combinators/exponential-families.jl
@@ -1,64 +1,105 @@
 export ExponentialFamily
 
-struct ExponentialFamily{B,Θ,H,T,A} <: AbstractKleisli
-    base::B
-    θ::Θ
-    η::H
-    t::T
-    a::A
+@concrete terse struct ExponentialFamily <: AbstractKleisli
+    base
+    mdim
+    pdim
+    t
+    x    
+    a
 end
 
-struct ExpFamMeasure{B,Θ,H,T,A,P} <: AbstractMeasure
-    base::B
-    θ::Θ
-    η::H
-    t::T
-    a::A
-    par :: P
+function ExponentialFamily(base, mdim, pdim, t, a)
+    return ExponentialFamily(base, mdim, pdim, t, I, a)
 end
-
-
-@inline function (fam::ExponentialFamily)(par)
-    base = fam.base
-    θ = fam.θ(par)
-    η = fam.η(θ)
-    t = fam.t
-    a = fam.a(θ)
-    ExpFamMeasure(base, θ, η, t, a, par)
-end
-
-basemeasure(d::ExpFamMeasure) = d.base
-
-function logdensity_def(d::ExpFamMeasure, x)
-    mydot(d.η, d.t(x)) - d.a
-end
-
 
 function MeasureBase.powermeasure(fam::ExponentialFamily, dims::NTuple{N,I}) where {N,I}
-    @inline t(x) = sum(fam.t, x)
-    a = AffineTransform((σ=prod(dims),)) ∘ fam.a
-    ExponentialFamily(fam.base ^ dims, fam.θ, fam.η, t, a)
+    t = Tuple((y -> f.(y) for f in fam.t))
+    a(η) = BroadcastArray(fam.a, η)
+    p = prod(dims)
+    ExponentialFamily(fam.base ^ dims, fam.mdim * p, fam.pdim * p,  t, fam.x, a)
 end
 
-struct ExpFamLikelihood{C,Θ,H,T,A} <: AbstractLikelihood
-    c::C
-    θ::Θ
-    η::H
-    t::T
-    a::A
+@concrete terse struct ExpFamMeasure <: AbstractMeasure
+    fam
+    η # instantiated to a value
+    a # instantiated to a value
+end
+
+
+@inline function (fam::ExponentialFamily)(β)
+    η = fam.x * β 
+    a = fam.a(η)
+    ExpFamMeasure(fam, η, a)
+end
+
+MeasureBase.basemeasure(d::ExpFamMeasure) = d.fam.base
+
+tracedot(a::AbstractVector,b::AbstractVector) = dot(a, b)
+
+tracedot(a::AbstractVector,x,b::AbstractVector) = dot(a, x, b)
+
+tracedot(a,b) = sum((dot(view(a,:,j), view(b,:,j)) for j in 1:size(a,2)))
+
+tracedot(a,x,b) = sum(1:size(a,2)) do j
+    dot(view(a,:,j), x, view(b,:,j))
+end
+
+# @inline function tracedot(a::BlockDiag, b::BlockDiag)
+#     numblocks = length(a.blocks)
+#     sum(tracedot(a.blocks[j], b.blocks[j]) for j in 1:length(a.blocks))
+# end
+
+# @inline function tracedot(a::BlockDiag, x::BlockDiag, b::BlockDiag)
+#     numblocks = length(x.blocks)
+#     sum(tracedot(a.blocks[j], x.blocks[j], b.blocks[j]) for j in 1:length(x.blocks))
+# end
+
+function logdensity_def(d::ExpFamMeasure, y)
[…]
