+++ src/utils.jl
@@ -1,4 +1,6 @@
 using LinearAlgebra
+export mydot
+using LazyArrays
 
 function solve(A::Union{AbstractMatrix, Factorization}, y::AbstractArray)
     (m, n) = size(A)
@@ -7,8 +9,29 @@ function solve(A::Union{AbstractMatrix, Factorization}, y::AbstractArray)
 end
 
 
+@inline function mydot(a::Real, b::Real)
+    return a * b
+end
+
+@inline function mydot(a::Ta, b::Tb) where {Ta<:AbstractArray, Tb<:AbstractArray}
+    return dot(a, b)
+end
+
+@inline function _mydot(a::NTuple{0}, b::NTuple{0}, s) 
+    return s
+end
+
+@inline function mydot(a::Tuple, b::Tuple) 
+    z = 0.0
+    _mydot(a, b, z)
+end
+
+@inline function _mydot(a::Tuple, b::Tuple, s)
+    _mydot(Base.tail(a), Base.tail(b), s + mydot(first(a), first(b)))
+end
+
 @generated function mydot(a::SVector{N,Ta}, b::SVector{N,Tb}) where {N,Ta,Tb}
-    z = zero(promote_type(Ta, Tb))
+    z = zero(float(Base.promote_eltype(Ta, Tb)))
     quote
         $(Expr(:meta, :inline))
         result = $z
