+++ src/transforms/ordered.jl
@@ -37,11 +37,11 @@ function TV.transform_with(flag::TV.LogJacFlag, t::Ordered, x, index::T) where {
 
     x = mappedarray(xj -> xj + OrderedΔx, x)
 
-    @inbounds (y[1], ℓ, _) = TV.transform_with(flag, as(Real, lo, hi), x, index)
+    @inbounds (y[1], ℓ, _) = TV.transform_with(flag, xform(Real, lo, hi), x, index)
     index += 1
 
     @inbounds for i in 2:len
-        (y[i], Δℓ, _) = TV.transform_with(flag, as(Real, y[i-1], hi), x, index)
+        (y[i], Δℓ, _) = TV.transform_with(flag, xform(Real, y[i-1], hi), x, index)
         ℓ = addlogjac(ℓ, Δℓ)
         index += 1
     end
@@ -74,7 +74,7 @@ end
 
 logdensity_def(s::Sorted, x) = logdensity_def(s.μ^s.n, x)
 
-TV.as(s::Sorted) = Ordered(as(s.μ), s.n)
+xform(s::Sorted) = Ordered(xform(s.μ), s.n)
 
 function Random.rand!(rng::AbstractRNG, d::Sorted, x::AbstractArray)
     rand!(rng, d.μ^d.n, x)
