+++ src/parameterized/uniform.jl
@@ -22,7 +22,7 @@ distproxy(::Uniform{()}) = Dists.Uniform()
 
 logdensity_def(d::Uniform{()}, x) = 0.0
 
-TV.as(::Uniform{()}) = as𝕀
+xform(::Uniform{()}) = as𝕀
 
 Base.rand(rng::AbstractRNG, T::Type, μ::Uniform{()}) = rand(rng, T)
 
