+++ src/parameterized/gumbel.jl
@@ -33,7 +33,7 @@ function Base.rand(rng::AbstractRNG, d::Gumbel{()})
     -log(-log(u))
 end
 
-TV.as(::Gumbel) = asℝ
+xform(::Gumbel) = asℝ
 
 ≪(::Gumbel, ::Lebesgue{X}) where {X<:Real} = true
 
