+++ src/smart-constructors.jl
@@ -5,7 +5,7 @@ affine(f::AffineTransform, μ::AbstractMeasure) = Affine(f, μ)
 
 affine(nt::NamedTuple, μ::AbstractMeasure) = affine(AffineTransform(nt), μ)
 
-affine(f) = μ -> affine(f, μ)
+affine(f) = Base.Fix1(affine, f)
 
 function affine(f::AffineTransform, parent::WeightedMeasure)
     WeightedMeasure(parent.logweight, affine(f, parent.base))
@@ -14,7 +14,6 @@ end
 function affine(f::AffineTransform, parent::FactoredBase)
     constℓ = parent.constℓ
     varℓ = parent.varℓ
-    # Avoid transforming `inbounds`, which is expensive
-    base = affine(f, restrict(parent.inbounds, parent.base))
-    FactoredBase(Returns(true), constℓ, varℓ, base)
+    base = affine(f, parent.base)
+    FactoredBase(constℓ, varℓ, base)
 end
