+++ src/parameterized/bernoulli.jl
@@ -9,9 +9,9 @@ import Base
 
 basemeasure(::Bernoulli) = Counting(Bool)
 
-@inline function logdensity_def(d::Bernoulli{(:p,)}, y)
+@inline function logdensity_def(d::Bernoulli{(:p,)}, y::Bool)
     p = d.p
-    return y * log(p) + (1 - y) * log(1 - p)
+    return ifelse(y, log(p), log(1 - p))
 end
 
 function density(d::Bernoulli{(:p,)}, y)
@@ -41,3 +41,5 @@ asparams(::Type{<:Bernoulli}, ::StaticSymbol{:logitp}) = asℝ
 
 distproxy(d::Bernoulli{(:p,)}) = Dists.Bernoulli(d.p)
 distproxy(d::Bernoulli{(:logitp,)}) = Dists.Bernoulli(logistic(d.logitp))
+
+insupport(::Bernoulli, x) = x ∈ (true, false)
