+++ src/parameterized/beta.jl
@@ -20,11 +20,10 @@ xform(::Beta) = as𝕀
 end
 
 @inline function basemeasure(d::Beta{(:α, :β)})
-    inbounds = in(𝕀)
     constℓ = 0.0
     varℓ = Returns(-logbeta(d.α, d.β))
     base = Lebesgue(ℝ)
-    FactoredBase(inbounds, constℓ, varℓ, base)
+    FactoredBase(constℓ, varℓ, base)
 end
 
 Base.rand(rng::AbstractRNG, T::Type, μ::Beta) = rand(rng, Dists.Beta(μ.α, μ.β))
@@ -33,3 +32,6 @@ distproxy(d::Beta{(:α, :β)}) = Dists.Beta(d.α, d.β)
 
 asparams(::Type{<:Beta}, ::StaticSymbol{:α}) = asℝ₊
 asparams(::Type{<:Beta}, ::StaticSymbol{:β}) = asℝ₊
+
+insupport(::Beta, x) = in𝕀(x)
+insupport(::Beta) = in𝕀
