### 2022-01-25 600fd17
condiional meaures
