### 2022-01-25 7885075
update affine smart constructor
