### 2022-01-27 cscherrer/MeasureTheory.jl#179 @cscherrer
Hi @ludgerpaehler , @mschauer and I were just discussing this. Most things under "Documentation" in the criteria seem to be about the repository, not the paper. Just to be sure we understand right, would adding this information in the documentation address your concern?
