+++ src/combinators/for.jl
@@ -1,0 +1,278 @@
+
+export For
+using Random
+import Base
+
+struct For{T, F, I} <: AbstractProductMeasure
+    f::F
+    inds::I
+
+    @inline function For{T}(f::F, inds::I) where {T,F,I<:Tuple}
+        new{T,instance_type(f),I}(f, inds)
+    end
+
+    @inline For{T,F,I}(f::F, inds::I) where {T,F,I} = new{T,F,I}(f,inds)
+
+    function For{Union{},F,I}(f::F, inds::I) where {F,I}
+        println("XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX")
+        println.(stacktrace())
+        println("XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX")
+        @show f(first(zip(inds...))...)
+        @error "Empty `For` construction"
+    end
+end
+
+
+@generated function For(f::F, inds::I) where {F,I<:Tuple}
+    eltypes = Tuple{eltype.(I.types)...}
+    quote
+        $(Expr(:meta, :inline))
+        T = Core.Compiler.return_type(f, $eltypes)
+        For{T,F,I}(f, inds)
+    end
+end
+
+
+# For(f, gen::Base.Generator) = ProductMeasure(Base.Generator(f ∘ gen.f, gen.iter))
+
+# @inline function tailcall_iter(f, iter)
+#     iter_result::Tuple = iterate(iter)
+#     (val, state) = iter_result
+#     tailcall_iter(f, iter, (val, state))
+# end
+
+# @inline function tailcall_iter(f, iter, ::Nothing)
+#     return unit(f)
+# end
+
+# unit(+) = false
+
+# @inline function tailcall_iter(f, iter, (val, state))
+#     tailcall_iter(f, iter, (val, state), unit(f))
+# end
+  
+# @inline function tailcall_iter(f, iter, (val, state), acc)
+#     tailcall_iter(f, iter, iterate(iter, state), f(val, acc))
+# end
+
+# @inline function tailcall_iter(f, iter, ::Nothing, acc)
+#     return acc
+# end
+
+# @inline function logdensity_def(d::For{T,F,I}, x::AbstractVector{X}; exec=SequentialEx(simd = true)) where {X,T,F,I<:Tuple{<:AbstractVector}}
+#     f(j) = @inbounds logdensity_def(d.f(j), x[j])
+#     f(j, ℓ) = ℓ + @inbounds logdensity_def(d.f(j), x[j])
+
+#     tailcall_iter(f, eachindex(x))
+# end
+
+@inline function logdensity_def(d::For{T,F,I}, x::AbstractVector{X}; exec=SequentialEx(simd = true)) where {X,T,F,I<:Tuple{<:AbstractVector}}
+    ind = only(d.inds)
+    @floop exec for j in eachindex(x)
+        local i = getindex(ind, j)
+        local Δℓ = @inbounds logdensity_def(d.f(i), x[j])
+        @reduce ℓ += Δℓ
+    end
+    ℓ
+end
+
+function logdensity_def(d::For, x::AbstractVector; exec=SequentialEx(simd = true)) 
+    @floop exec for j in eachindex(x)
+        local i = (getindex(ind, j) for ind in d.inds)
+        local Δℓ = @inbounds logdensity_def(d.f(i...), x[j])
+        @reduce ℓ += Δℓ
+    end
+ 
+    ℓ
+end
+   
+
[…]
