+++ src/utils.jl
@@ -44,3 +44,63 @@ end
 
 in𝕀(x) = static(0.0) ≤ x ≤ static(1.0)
 inℝ₊(x) = static(0.0) ≤ x
+
+
+_eltype(T) = eltype(T)
+
+function _eltype(g::Base.Generator{I}) where {I}
+    Core.Compiler.return_type(g.f, Tuple{_eltype(I)})
+end
+
+function _eltype(::Type{Base.Generator{I,ComposedFunction{Outer,Inner}}}) where {Outer,Inner,I}
+    _eltype(Base.Generator{_eltype(Base.Generator{I,Inner}), Outer})
+end
+
+function _eltype(::Type{Base.Generator{I,F}}) where {F<:Function,I}
+    f = instance(F)
+    Core.Compiler.return_type(f, Tuple{_eltype(I)})
+end
+
+function _eltype(::Type{Z}) where {Z<:Iterators.Zip}
+    map(_eltype, Z.types[1].types)
+end
+
+
+
+# Adapted from https://github.com/JuliaArrays/MappedArrays.jl/blob/46bf47f3388d011419fe43404214c1b7a44a49cc/src/MappedArrays.jl#L229
+function func_string(f, types)
+    ft = typeof(f)
+    mt = ft.name.mt
+    name = string(mt.name)
+    if startswith(name, '#')
+        # This is an anonymous function. See if it can be printed nicely
+        lwrds = code_lowered(f, types)
+        if length(lwrds) != 1
+            return string(f)
+        end
+        lwrd = lwrds[1]
+        c = lwrd.code
+        if length(c) == 2 && ((isa(c[2], Expr) && c[2].head === :return) || (isdefined(Core, :ReturnNode) && isa(c[2], Core.ReturnNode)))
+            # This is a single-line anonymous function, we should handle it
+            s = lwrd.slotnames[2:end]
+            result = ""
+            if length(s) == 1
+                result *= string(s[1])
+                result *= "->"
+            else
+                result *= string(tuple(s...))
+                result *= "->"
+            end
+            c1 = string(c[1])
+            for i = 1:length(s)
+                c1 = replace(c1, "_"*string(i+1)=>string(s[i]))
+            end
+            result *= c1
+            return result
+        else
+            return string(f)
+        end
+    else
+        return string(f)
+    end
+end
