+++ jet-test.jl
@@ -1,0 +1,5 @@
+using MeasureTheory
+
+dist = For(3) do j Normal(σ=j) end
+
+b = basemeasure_depth(dist)
