### 2022-02-08 cscherrer/MeasureTheory.jl#179 @ludgerpaehler
@cscherrer sorry for the belated reply, yes, absolutely! My concerns were mostly about the docs :)
