### 2022-02-11 8b1e845
CompatHelper: bump compat for TransformVariables to 0.6, (keep existing compat)
