+++ Project.toml
@@ -36,7 +36,7 @@ ConcreteStructs = "0.2"
 ConstructionBase = "1.3"
 Distributions = "0.25"
 DynamicIterators = "0.4"
-FillArrays = "0.12"
+FillArrays = "0.12, 0.13"
 InfiniteArrays = "0.11, 0.12"
 KeywordCalls = "0.2"
 LogExpFunctions = "0.3.3"
