### 2022-02-16 a034cfb
Make Bernoulli more efficient
