+++ Project.toml
@@ -44,7 +44,7 @@ MLStyle = "0.4"
 MacroTools = "0.5"
 MappedArrays = "0.4"
 MeasureBase = "0.5"
-NamedTupleTools = "0.13"
+NamedTupleTools = "0.13, 0.14"
 NestedTuples = "0.3"
 PositiveFactorizations = "0.2"
 PrettyPrinting = "0.3"
