+++ Project.toml
@@ -48,7 +48,7 @@ DensityInterface = "0.4"
 Distributions = "0.25"
 DynamicIterators = "0.4"
 FLoops = "0.2"
-FillArrays = "0.12"
+FillArrays = "0.12, 0.13"
 Infinities = "0.1"
 InverseFunctions = "0.1"
 KeywordCalls = "0.2"
@@ -57,17 +57,17 @@ LogExpFunctions = "0.3.3"
 MLStyle = "0.4"
 MacroTools = "0.5"
 MappedArrays = "0.4"
-MeasureBase = "0.5"
-NamedTupleTools = "0.13"
+MeasureBase = "0.6"
+NamedTupleTools = "0.13, 0.14"
 NestedTuples = "0.3"
 PositiveFactorizations = "0.2"
-PrettyPrinting = "0.3"
+PrettyPrinting = "0.3, 0.4"
 Reexport = "1"
 SpecialFunctions = "1, 2"
-Static = "0.4, 0.5"
+Static = "0.5, 0.6"
 StaticArrays = "1.3"
 StatsFuns = "0.9"
-TransformVariables = "0.4, 0.5"
+TransformVariables = "0.5, 0.6"
 Tricks = "0.1"
 julia = "1.5"
 
