+++ src/parameterized/cauchy.jl
@@ -3,48 +3,49 @@
 
 export Cauchy, HalfCauchy
 
-@parameterized Cauchy(μ,σ) ≃ (1/π) * Lebesgue(ℝ)
+@parameterized Cauchy(μ, σ)
 
 @kwstruct Cauchy()
 @kwstruct Cauchy(μ)
 @kwstruct Cauchy(σ)
-@kwstruct Cauchy(μ,σ)
+@kwstruct Cauchy(μ, σ)
 @kwstruct Cauchy(ω)
-@kwstruct Cauchy(μ,ω)
+@kwstruct Cauchy(μ, ω)
 
 
+logdensity_def(d::Cauchy, x) = logdensity_def(proxy(d), x)
 
-for N in AFFINEPARS
-    @eval begin
-        proxy(d::Cauchy{$N}) = affine(params(d), Cauchy())
-        logdensity(d::Cauchy{$N}, x) = logdensity(proxy(d), x)
-        basemeasure(d::Cauchy{$N}) = basemeasure(proxy(d))
-    end
-end
+basemeasure(d::Cauchy) = WeightedMeasure(static(-logπ), Lebesgue(ℝ))
 
 # @affinepars Cauchy
 
-function logdensity(d::Cauchy{()} , x) 
+@inline function logdensity_def(d::Cauchy{()}, x)
     return -log1p(x^2)
 end
 
-function density(d::Cauchy{()} , x) 
+function density_def(d::Cauchy{()}, x)
     return inv(1 + x^2)
 end
 
 Base.rand(rng::AbstractRNG, T::Type, μ::Cauchy{()}) = randn(rng, T) / randn(rng, T)
 
-≪(::Cauchy, ::Lebesgue{X}) where X <: Real = true
+for N in AFFINEPARS
+    @eval begin
+        proxy(d::Cauchy{$N}) = affine(params(d), Cauchy())
+        @useproxy Cauchy{$N}
+        rand(rng::AbstractRNG, ::Type{T}, d::Cauchy{$N}) where {T} = rand(rng, T, affine(params(d), Cauchy()))
+    end
+end
 
-TV.as(::Cauchy) = asℝ
+
+≪(::Cauchy, ::Lebesgue{X}) where {X<:Real} = true
+
+xform(::Cauchy) = asℝ
 
 @half Cauchy
 
-HalfCauchy(σ) = HalfCauchy(σ=σ)
+HalfCauchy(σ) = HalfCauchy(σ = σ)
 
 distproxy(d::Cauchy{()}) = Dists.Cauchy()
-distproxy(d::Cauchy{(:μ,)}) = Dists.Cauchy(d.μ, 1.0)
-distproxy(d::Cauchy{(:σ,)}) = Dists.Cauchy(0.0, d.σ)
-distproxy(d::Cauchy{(:μ,:σ)}) = Dists.Cauchy(d.μ, d.σ)
-distproxy(d::Cauchy{(:ω,)}) = Dists.Cauchy(0.0, inv(d.ω))
-distproxy(d::Cauchy{(:μ,:ω)}) = Dists.Cauchy(d.μ, inv(d.ω))
+
+insupport(::Cauchy, x) = true
