+++ src/parameterized/uniform.jl
@@ -6,23 +6,25 @@ export Uniform
 @parameterized Uniform()
 @kwstruct Uniform()
 
-
 ###############################################################################
 # Standard Uniform
 
-function basemeasure(::Uniform{()})
-    inbounds(x) = 0 < x < 1
-    constℓ = 0.0
-    varℓ() = 0.0
+
+insupport(::Uniform{()}) = in𝕀
+insupport(::Uniform{()}, x) = in𝕀(x)
+
+@inline function basemeasure(::Uniform{()})
+    constℓ = static(0.0)
+    varℓ = Returns(static(0.0))
     base = Lebesgue(ℝ)
-    FactoredBase(inbounds, constℓ, varℓ, base)
+    FactoredBase(constℓ, varℓ, base)
 end
 
 distproxy(::Uniform{()}) = Dists.Uniform()
 
-logdensity(d::Uniform{()}, x) = 0.0
+logdensity_def(d::Uniform{()}, x) = 0.0
 
-TV.as(::Uniform{()}) = as𝕀
+xform(::Uniform{()}) = as𝕀
 
 Base.rand(rng::AbstractRNG, T::Type, μ::Uniform{()}) = rand(rng, T)
 
