+++ src/parameterized/dirichlet.jl
@@ -1,26 +1,31 @@
 # Dirichlet distribution
+using MeasureBase: Simplex
 
 export Dirichlet
 
+using FillArrays
+
 @parameterized Dirichlet(α)
 
-TV.as(d::Dirichlet{(:α,)}) = TV.UnitSimplex(length(d.α))
+xform(d::Dirichlet{(:α,)}) = TV.UnitSimplex(length(d.α))
 
-function basemeasure(μ::Dirichlet{(:α,)})
-    t = as(μ)
-    d = dimension(t)
-    return Pushforward(t, Lebesgue(ℝ)^d, false)
+@inline function basemeasure(μ::Dirichlet{(:α,)})
+    α = μ.α
+    t = xform(μ)
+    d = TV.dimension(t)
+    logw = loggamma(sum(α)) - sum(loggamma, α)
+    return WeightedMeasure(logw, Lebesgue(Simplex()))
 end
 
 @kwstruct Dirichlet(α)
 
 Dirichlet(k::Integer, α) = Dirichlet(Fill(α, k))
 
-function logdensity(d::Dirichlet{(:α,)}, x)
+@inline function logdensity_def(d::Dirichlet{(:α,)}, x)
     α = d.α
     s = 0.0
     for j in eachindex(x)
-        s += xlogy(α[j], x[j])
+        s += xlogy(α[j] - 1, x[j])
     end
     return s
 end
@@ -31,5 +36,9 @@ distproxy(d::Dirichlet{(:α,)}) = Dists.Dirichlet(d.α)
 
 function testvalue(d::Dirichlet{(:α,)})
     n = length(d.α)
-    Fill(1/n, n)
+    Fill(1 / n, n)
+end
+
+@inline function insupport(d::Dirichlet{(:α,)}, x) 
+    length(x) == length(d.α) && sum(x) ≈ 1.0 
 end

[…]
