+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.14",
   "v0.13",
   "v0.12",
   "v0.11",
@@ -14,5 +15,5 @@ var DOC_VERSIONS = [
   "v0.2",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.13.2";
+var DOCUMENTER_NEWEST = "v0.14.0";
 var DOCUMENTER_STABLE = "stable";
