+++ src/parameterized/dirichlet.jl
@@ -7,7 +7,7 @@ using FillArrays
 
 @parameterized Dirichlet(α)
 
-xform(d::Dirichlet{(:α,)}) = TV.UnitSimplex(length(d.α))
+TV.as(d::Dirichlet{(:α,)}) = TV.UnitSimplex(length(d.α))
 
 @inline function basemeasure(μ::Dirichlet{(:α,)})
     α = μ.α
