### 2022-04-05 695bec6
enumerate instead of zip
