### 2022-04-05 0be2343
drop `static` from `insupport` results
