+++ src/parameterized/bernoulli.jl
@@ -40,7 +40,7 @@ Base.rand(rng::AbstractRNG, T::Type, d::Bernoulli{(:logitp,)}) =
 asparams(::Type{<:Bernoulli}, ::StaticSymbol{:p}) = as𝕀
 asparams(::Type{<:Bernoulli}, ::StaticSymbol{:logitp}) = asℝ
 
-distproxy(d::Bernoulli{(:p,)}) = Dists.Bernoulli(d.p)
-distproxy(d::Bernoulli{(:logitp,)}) = Dists.Bernoulli(logistic(d.logitp))
+proxy(d::Bernoulli{(:p,)}) = Dists.Bernoulli(d.p)
+proxy(d::Bernoulli{(:logitp,)}) = Dists.Bernoulli(logistic(d.logitp))
 
 insupport(::Bernoulli, x) = x ∈ (true, false)
