+++ src/parameterized/beta.jl
@@ -28,7 +28,7 @@ end
 
 Base.rand(rng::AbstractRNG, T::Type, μ::Beta) = rand(rng, Dists.Beta(μ.α, μ.β))
 
-distproxy(d::Beta{(:α, :β)}) = Dists.Beta(d.α, d.β)
+proxy(d::Beta{(:α, :β)}) = Dists.Beta(d.α, d.β)
 
 asparams(::Type{<:Beta}, ::StaticSymbol{:α}) = asℝ₊
 asparams(::Type{<:Beta}, ::StaticSymbol{:β}) = asℝ₊
