+++ src/parameterized/multinomial.jl
@@ -27,7 +27,7 @@ end
 Base.rand(rng::AbstractRNG, T::Type, μ::Multinomial) =
     rand(rng, Dists.Multinomial(μ.n, μ.p))
 
-distproxy(d::Multinomial{(:p,)}) = Dists.Multinomial(d.n, d.p)
+proxy(d::Multinomial{(:p,)}) = Dists.Multinomial(d.n, d.p)
 
 # Based on
 # https://github.com/JuliaMath/Combinatorics.jl/blob/c2114a71ccfc93052efb9a9379e62b81b9388ef8/src/factorials.jl#L99
