+++ src/parameterized/cauchy.jl
@@ -46,6 +46,6 @@ TV.as(::Cauchy) = asℝ
 
 HalfCauchy(σ) = HalfCauchy(σ = σ)
 
-distproxy(d::Cauchy{()}) = Dists.Cauchy()
+proxy(d::Cauchy{()}) = Dists.Cauchy()
 
 insupport(::Cauchy, x) = true
