+++ src/parameterized/studentt.jl
@@ -59,12 +59,12 @@ xform(::StudentT) = asℝ
 
 Base.rand(rng::AbstractRNG, T::Type, μ::StudentT{(:ν,)}) = rand(rng, Dists.TDist(μ.ν))
 
-distproxy(d::StudentT{(:ν,)}) = Dists.TDist(d.ν)
-distproxy(d::StudentT{(:ν, :μ)}) = Dists.LocationScale(d.μ, 1.0, Dists.TDist(d.ν))
-distproxy(d::StudentT{(:ν, :σ)}) = Dists.LocationScale(0.0, d.σ, Dists.TDist(d.ν))
-distproxy(d::StudentT{(:ν, :ω)}) = Dists.LocationScale(0.0, inv(d.ω), Dists.TDist(d.ν))
-distproxy(d::StudentT{(:ν, :μ, :σ)}) = Dists.LocationScale(d.μ, d.σ, Dists.TDist(d.ν))
-distproxy(d::StudentT{(:ν, :μ, :ω)}) = Dists.LocationScale(d.μ, inv(d.ω), Dists.TDist(d.ν))
+proxy(d::StudentT{(:ν,)}) = Dists.TDist(d.ν)
+proxy(d::StudentT{(:ν, :μ)}) = affine((μ=d.μ,), StudentT(d.ν))
+proxy(d::StudentT{(:ν, :σ)}) = affine((σ=d.σ,), StudentT(d.ν))
+proxy(d::StudentT{(:ν, :ω)}) = affine((ω=d.ω,), StudentT(d.ν))
+proxy(d::StudentT{(:ν, :μ, :σ)}) = affine((μ= d.μ, σ = d.σ), StudentT(d.ν))
+proxy(d::StudentT{(:ν, :μ, :ω)}) = affine((μ= d.μ, ω = d.ω), StudentT(d.ν))
 
 @half StudentT
 
