### 2022-04-05 7b36be7
IfElse.ifelse(p::Bernoulli, t, f)
