### 2022-04-07 e2f3db1
update some base measures
