### 2022-04-08 4be9c3f
fix some redundancies
