+++ src/parameterized/exponential.jl
@@ -89,5 +89,3 @@ end
 proxy(d::Exponential{(:logλ,)}) = Dists.Exponential(exp(-d.logλ))
 
 asparams(::Type{<:Exponential}, ::StaticSymbol{:logλ}) = asℝ
-
-insupport(::Exponential, x) = x ≥ 0
