### 2022-04-10 a822cfc
instance_type => Core.Typeof
