### 2022-04-11 e974353
update testvalue for Bernoulli and Binomial
