+++ test/runtests.jl
@@ -407,7 +407,7 @@ end
 @testset "Density measures and Radon-Nikodym" begin
     x = randn()
     let d = ∫(𝒹(Cauchy(), Normal()), Normal())
-        @test_broken logdensityof(𝒹(d, Cauchy()), x) ≈ 0 atol=1e-12
+        @test logdensityof(𝒹(d, Cauchy()), x) ≈ 0 atol=1e-12
     end
 
     let f = 𝒹(∫(x -> x^2, Normal()), Normal())
