### 2022-04-11 71f097e
Fall-back `For` method for when inference fails
