+++ Project.toml
@@ -59,7 +59,7 @@ LogExpFunctions = "0.3.3"
 MLStyle = "0.4"
 MacroTools = "0.5"
 MappedArrays = "0.4"
-MeasureBase = "0.6"
+MeasureBase = "0.7"
 NamedTupleTools = "0.13, 0.14"
 NestedTuples = "0.3"
 PositiveFactorizations = "0.2"
