+++ test/runtests.jl
@@ -565,4 +565,4 @@ end
     x = randn()
     @test logdensityof(ifelse(Bernoulli(p), Normal(), Normal()), x) == logdensityof(Normal(), x)
     @test logdensityof(ifelse(Bernoulli(p), Normal(2,3), Normal()), x) == logdensityof(p*Normal(2,3) + (1-p) * Normal(), x)
-end 
+end
