### 2022-04-13 ec640fc
add Julia 1.7 to CI
