+++ Project.toml
@@ -71,7 +71,7 @@ StaticArrays = "1.3"
 StatsFuns = "0.9"
 TransformVariables = "0.5, 0.6"
 Tricks = "0.1"
-julia = "1.5"
+julia = "1.6"
 
 [extras]
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
