### 2022-04-14 43f60f3
CompatHelper: bump compat for MeasureBase to 0.7, (keep existing compat)
