### 2022-04-16 3998ca5
drop 1.7 CI (seems buggy?)
