+++ src/parameterized.jl
@@ -74,4 +74,4 @@ function asparams(μ::M, nt::NamedTuple = NamedTuple()) where {M<:ParameterizedM
     asparams(constructorof(M), nt)
 end
 
-xform(::Half) = asℝ₊
+TV.as(::Half) = asℝ₊
