+++ src/parameterized/cauchy.jl
@@ -40,12 +40,12 @@ end
 
 ≪(::Cauchy, ::Lebesgue{X}) where {X<:Real} = true
 
-xform(::Cauchy) = asℝ
+TV.as(::Cauchy) = asℝ
 
 @half Cauchy
 
 HalfCauchy(σ) = HalfCauchy(σ = σ)
 
-distproxy(d::Cauchy{()}) = Dists.Cauchy()
+proxy(d::Cauchy{()}) = Dists.Cauchy()
 
 insupport(::Cauchy, x) = true

[…]
