+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.15",
   "v0.14",
   "v0.13",
   "v0.12",
@@ -15,5 +16,5 @@ var DOC_VERSIONS = [
   "v0.2",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.14.0";
+var DOCUMENTER_NEWEST = "v0.15.0";
 var DOCUMENTER_STABLE = "stable";
