### 2022-04-18 ab96459
fix MvNormal bug (#189)
* fix MvNormal bug

* bump version
