+++ versions.js
@@ -16,5 +16,5 @@ var DOC_VERSIONS = [
   "v0.2",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.15.0";
+var DOCUMENTER_NEWEST = "v0.15.1";
 var DOCUMENTER_STABLE = "stable";
