### 2022-04-20 862c0c0
entropy
