### 2022-04-27 eb3656b
more gamma stuff
