+++ src/parameterized/gamma.jl
@@ -3,28 +3,35 @@
 
 export Gamma
 
-@parameterized Gamma(shape) ≃ Lebesgue(ℝ₊)
+@parameterized Gamma(k)
 
-@inline function logdensity_def(μ::Gamma{(:shape,)}, x)
-    α = μ.shape
-    xinv = 1 / x
+@kwstruct Gamma(k)
+@kwstruct Gamma(k, θ)
 
-    return xlogy(α + 1, xinv) - xinv - loggamma(α)
+Gamma(k,θ) = Gamma((k=k, θ=θ))
+
+@useproxy Gamma{(:k, :θ)}
+
+@inline function logdensity_def(d::Gamma{(:k,)}, x)
+    return xlogy(d.k - 1, x) - x
+end
+
+function basemeasure(d::Gamma{(:k,)})
+    constℓ = 0.0
+    varℓ() = -loggamma(d.k)
+    base = LebesgueMeasure()
+    FactoredBase(constℓ, varℓ, base)
+end
+
+function proxy(d::Gamma{(:k, :θ)})
+    affine(NamedTuple{(:σ,)}(d.θ), Gamma((k = d.k,)))
 end
 
 
-for N in [(:σ,), (:ω,)]
-    @eval begin
-        proxy(d::Gamma{(:shape, $N...)}) =
-            affine(NamedTuple{$N}(params(d)), Gamma((shape = d.shape,)))
-    end
-end
-
-
-Base.rand(rng::AbstractRNG, T::Type, μ::Gamma{(:shape,)}) =
-    rand(rng, Dists.Gamma(μ.shape))
+Base.rand(rng::AbstractRNG, T::Type, μ::Gamma{(:k,)}) =
+    rand(rng, Dists.Gamma(μ.k))
 
 
 TV.as(::Gamma) = asℝ₊
 
-# @μσ_methods Gamma(shape)
+insupport(::Gamma, x) = x > 0
