### 2022-04-27 79922fc
Beroulli()
