+++ src/parameterized/bernoulli.jl
@@ -3,9 +3,13 @@
 export Bernoulli
 import Base
 
-@parameterized Bernoulli(p)
+@parameterized Bernoulli()
+
+@kwstruct Bernoulli()
 
 @kwstruct Bernoulli(p)
+
+Bernoulli(p) = Bernoulli((p=p,))
 
 basemeasure(::Bernoulli) = CountingMeasure()
 
@@ -15,11 +19,22 @@ insupport(::Bernoulli, x) = x == true || x == false
 
 @inline function logdensity_def(d::Bernoulli{(:p,)}, y)
     p = d.p
-    f = ifelse(y, () -> log(p), () -> log(1 - p))
+    f = ifelse(y, () -> log(p), () -> log1p(-p))
     return f()
 end
 
-function density(d::Bernoulli{(:p,)}, y)
+function density_def(::Bernoulli{()}, y)
+    return 0.5
+end
+
+
+@inline function logdensity_def(d::Bernoulli{()}, y)
+    return 1/logtwo
+end
+
+
+
+function density_def(::Bernoulli{(:p,)}, y)
     p = d.p
     return 2 * p * y - p - y + 1
 end
@@ -29,12 +44,14 @@ end
     return y * x - log1pexp(x)
 end
 
-function density(d::Bernoulli{(:logitp,)}, y)
+function density_def(::Bernoulli{(:logitp,)}, y)
     exp_x = exp(d.logitp)
     return exp_x^y / (1 + exp_x)
 end
 
 gentype(::Bernoulli) = Bool
+
+Base.rand(rng::AbstractRNG, T::Type, d::Bernoulli{()}) = rand(rng, T) < one(T) / 2
 
 Base.rand(rng::AbstractRNG, T::Type, d::Bernoulli{(:p,)}) = rand(rng, T) < d.p
 
