### 2022-04-27 bd3a325
inverse Gaussian
