+++ src/parameterized/bernoulli.jl
@@ -19,8 +19,7 @@ insupport(::Bernoulli, x) = x == true || x == false
 
 @inline function logdensity_def(d::Bernoulli{(:p,)}, y)
     p = d.p
-    f = ifelse(y, () -> log(p), () -> log1p(-p))
-    return f()
+    y == true ? log(p) : log1p(-p)
 end
 
 function density_def(::Bernoulli{()}, y)
@@ -29,12 +28,11 @@ end
 
 
 @inline function logdensity_def(d::Bernoulli{()}, y)
-    return 1/logtwo
+    return -logtwo
 end
 
 
-
-function density_def(::Bernoulli{(:p,)}, y)
+function density_def(d::Bernoulli{(:p,)}, y)
     p = d.p
     return 2 * p * y - p - y + 1
 end
@@ -44,7 +42,7 @@ end
     return y * x - log1pexp(x)
 end
 
-function density_def(::Bernoulli{(:logitp,)}, y)
+function density_def(d::Bernoulli{(:logitp,)}, y)
     exp_x = exp(d.logitp)
     return exp_x^y / (1 + exp_x)
 end
