+++ src/combinators/tweedie.jl
@@ -1,37 +1,62 @@
 export Tweedie
 
-struct Tweedie{B,Θ,P} <: AbstractKleisli
+abstract type AbstractEDM <: AbstractKleisli end
+
+
+"""
+https://en.wikipedia.org/wiki/Tweedie_distribution
+
+The Tweedie distributions include a number of familiar distributions as well as
+some unusual ones, each being specified by the domain of the index parameter. We
+have the
+
+    extreme stable distribution, p < 0,
+    normal distribution, p = 0,
+    Poisson distribution, p = 1,
+    compound Poisson-gamma distribution, 1 < p < 2,
+    gamma distribution, p = 2,
+    positive stable distributions, 2 < p < 3,
+    Inverse Gaussian distribution, p = 3,
+    positive stable distributions, p > 3, and
+    extreme stable distributions, p = ∞.
+
+For 0 < p < 1 no Tweedie model exists. Note that all stable distributions mean
+actually generated by stable distributions. 
+"""
+struct Tweedie{S,B,D,P} <: AbstractEDM
+    support_contains::S
     base::B
-    θ::Θ
+    dim::D
     p::P
 end
 
 struct TweedieMeasure{B,Θ,P,S,C} <: AbstractMeasure
-    base::B
+    fam::Tweedie{B,Θ,P}
     θ::Θ
-    p::P
     σ::S
     cumulant::C
 end
 
-@inline function tweedie_mean(::StaticFloat64{0.0}, θ)
-    return θ
-end
 
-@inline function tweedie_mean(::StaticFloat64{1.0}, θ)
-    return exp(θ)
-end
 
-@inline function tweedie_mean(::StaticFloat64{2.0}, θ)
-    return inv(log(-θ))
-end
+mean(d::TweedieMeasure) = tweedie_mean(d.fam.p, d.θ)
 
-@generated function tweedie_mean(::StaticFloat64{p}, θ) where p
-    α_minus_1 = (p-2) / (p-1) - 1
+var(d::TweedieMeasure) = d.σ^2 * mean(d) ^ d.fam.p
 
-    quote
-        $(Expr(:meta, :inline))
-        (θ / α_minus_1) ^ α_minus_1
+###############################################################################
+# Tweedie cumulants
+
+@inline function tweedie_cumulant(p::P, θ) where {P}
+    if p == zero(P)
+        return 0.5 * θ ^ 2
+    elseif p == one(P)
+        return exp(θ)
+    elseif p == 2
+        return -log(-θ)
+    else
+        α = (p-2) / (p-1)
+        coeff = (α-1) / α
+        return coeff * (θ/(α-1)) ^ α
     end
 end
 
@@ -65,6 +90,44 @@ end
     a = fam.a(θ)
     TweedieMeasure(base, θ, p, σ)
 end
+
+###############################################################################
+# Tweedie mean function
+
+@inline function tweedie_mean(p::P, θ) where {P}
+    if p == zero(P)
+        return θ
+    elseif p == one(P)
+        return exp(θ)
+    elseif p == 2
+        return inv(log(-θ))
+    else
+        α_minus_1 = (p-2) / (p-1) - 1
+        return (θ / α_minus_1) ^ α_minus_1
+    end
+end
+
+@inline function tweedie_mean(::StaticFloat64{0.0}, θ)
+    return θ
+end
+
+@inline function tweedie_mean(::StaticFloat64{1.0}, θ)
+    return exp(θ)
+end
+
+@inline function tweedie_mean(::StaticFloat64{2.0}, θ)
+    return inv(log(-θ))
+end
+
[…]
