+++ src/parameterized/poisson.jl
@@ -2,23 +2,30 @@
 
 export Poisson
 import Base
-using SpecialFunctions: logfactorial
+using SpecialFunctions: loggamma
 
-@parameterized Poisson(λ)
+@parameterized Poisson()
 
+@kwstruct Poisson()
 @kwstruct Poisson(λ)
+
+Poisson(λ) = Poisson((λ=λ,))
 
 basemeasure(::Poisson) = Counting(BoundedInts(static(0), static(Inf)))
 
+@inline function logdensity_def(d::Poisson{()}, y)
+    return y - λ - loggamma(1 + y)
+end
+
 @inline function logdensity_def(d::Poisson{(:λ,)}, y)
     λ = d.λ
-    return y * log(λ) - λ - logfactorial(y)
+    return y * log(λ) - λ - loggamma(1 + y)
 end
 
 @kwstruct Poisson(logλ)
 
 @inline function logdensity_def(d::Poisson{(:logλ,)}, y)
-    return y * d.logλ - exp(d.logλ) - logfactorial(y)
+    return y * d.logλ - exp(d.logλ) - loggamma(1 + y)
 end
 
 asparams(::Type{<:Poisson}, ::StaticSymbol{:λ}) = asℝ₊
