### 2022-04-28 70044f8
drop pdf and logpdf
