### 2022-04-28 a544dcd
update Normal(μ, σ²)
