### 2022-04-28 0e446a1
Gamma(μ, ϕ) for GLMs
