+++ src/parameterized/normal.jl
@@ -63,6 +63,7 @@ TV.as(::Normal) = asℝ
     std   => σ
     sigma => σ
     var   => σ²
+    ϕ     => σ²
 ]
 
 # It's often useful to be able to map into the parameter space for a given
