### 2022-04-30 58c3392
CompatHelper: bump compat for StatsFuns to 1, (keep existing compat)
