+++ Project.toml
@@ -68,7 +68,7 @@ Reexport = "1"
 SpecialFunctions = "1, 2"
 Static = "0.5, 0.6"
 StaticArrays = "1.3"
-StatsFuns = "0.9"
+StatsFuns = "0.9, 1"
 TransformVariables = "0.5, 0.6"
 Tricks = "0.1"
 julia = "1.6"
