### 2022-05-02 3c7bef3
drop FactoredBase
