+++ src/smart-constructors.jl
@@ -11,13 +11,6 @@ function affine(f::AffineTransform, parent::WeightedMeasure)
     WeightedMeasure(parent.logweight, affine(f, parent.base))
 end
 
-function affine(f::AffineTransform, parent::FactoredBase)
-    constℓ = parent.constℓ
-    varℓ = parent.varℓ
-    base = affine(f, parent.base)
-    FactoredBase(constℓ, varℓ, base)
-end
-
 using MeasureBase: RealNumbers
 
 function affine(f::AffineTransform{(:μ, :σ)}, parent::Lebesgue{RealNumbers})
