### 2022-05-02 bed2f0c
drop old LazyArrays dependency
