+++ src/distributions.jl
@@ -41,3 +41,5 @@ logdensity_def(μ::Dists.Distribution, x) = Dists.logpdf(μ, x)
 density_def(μ::Dists.Distribution, x) = Dists.pdf(μ, x)
 
 testvalue(d::Dists.Distribution) = rand(d)
+
+insupport(d::Distribution, x) = logdensityof(d, x) > -Inf
