+++ src/distributions.jl
@@ -42,4 +42,4 @@ density_def(μ::Dists.Distribution, x) = Dists.pdf(μ, x)
 
 testvalue(d::Dists.Distribution) = rand(d)
 
-insupport(d::Distribution, x) = logdensityof(d, x) > -Inf
+insupport(d::Dists.Distribution, x) = logdensityof(d, x) > -Inf
