### 2022-05-04 ead400c
don't export `ifelse` (https://github.com/cscherrer/MeasureTheory.jl/issues/192)
