+++ src/MeasureTheory.jl
@@ -45,7 +45,6 @@ import LogExpFunctions
 import NamedTupleTools
 import InverseFunctions: inverse
 export inverse
-export ifelse
 
 import MeasureBase: insupport, instance, marginals
 import MeasureBase:
