### 2022-05-04 17a2cf4
Kleisli => TransitionKernel
