### 2022-05-05 be07168
work on truncated and censored
