### 2022-05-09 439bf9e
improve func_string
