+++ src/utils.jl
@@ -87,7 +87,7 @@ function func_string(f, types)
                 result *= string(s[1])
                 result *= "->"
             else
-                result *= string(tuple(s...))
+                result *= "(" * join(string.(tuple(s...)), ", ") * ")"
                 result *= "->"
             end
             c1 = string(c[1])
