### 2022-05-09 52b0a7a
Simplify logdensity_def(::For, x)
