+++ Project.toml
@@ -12,7 +12,6 @@ ConstructionBase = "187b0558-2788-49d3-abe0-74a17ed4e7c9"
 DensityInterface = "b429d917-457f-4dbc-8f4c-0cc954292b1d"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 DynamicIterators = "6c76993d-992e-5bf1-9e63-34920a5a5a38"
-FLoops = "cc61a311-1640-44b5-9fba-1b764f453329"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 IfElse = "615f187c-cbe4-4ef1-ba3b-2fcf58d6d173"
 Infinities = "e1ba4f0e-776d-440f-acd9-e1d2e9742647"
@@ -49,7 +48,6 @@ ConstructionBase = "1.3"
 DensityInterface = "0.4"
 Distributions = "0.25"
 DynamicIterators = "0.4"
-FLoops = "0.2"
 FillArrays = "0.12, 0.13"
 IfElse = "0.1"
 Infinities = "0.1"
