+++ src/combinators/for.jl
@@ -3,7 +3,7 @@ export For
 using Random
 import Base
 
-struct For{T, F, I} <: AbstractProductMeasure
+struct For{T,F,I} <: AbstractProductMeasure
     f::F
     inds::I
 
@@ -11,7 +11,7 @@ struct For{T, F, I} <: AbstractProductMeasure
         new{T,Core.Typeof(f),I}(f, inds)
     end
 
-    @inline For{T,F,I}(f::F, inds::I) where {T,F,I} = new{T,F,I}(f,inds)
+    @inline For{T,F,I}(f::F, inds::I) where {T,F,I} = new{T,F,I}(f, inds)
 
     function For{Union{},F,I}(f::F, inds::I) where {F,I}
         println("XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX")
@@ -20,10 +20,9 @@ struct For{T, F, I} <: AbstractProductMeasure
         println.(stacktrace())
         println("XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX")
         # @error "Empty `For` construction"
-        return ProductMeasure(mappedarray(i -> f(Tuple(i)...), CartesianIndices(inds...))) 
+        return ProductMeasure(mappedarray(i -> f(Tuple(i)...), CartesianIndices(inds...)))
     end
 end
-
 
 @generated function For(f::F, inds::I) where {F,I<:Tuple}
     eltypes = Tuple{eltype.(I.types)...}
@@ -33,7 +32,6 @@ end
         For{T,F,I}(f, inds)
     end
 end
-
 
 # For(f, gen::Base.Generator) = ProductMeasure(Base.Generator(f ∘ gen.f, gen.iter))
 
@@ -52,7 +50,7 @@ end
 # @inline function tailcall_iter(f, iter, (val, state))
 #     tailcall_iter(f, iter, (val, state), unit(f))
 # end
-  
+
 # @inline function tailcall_iter(f, iter, (val, state), acc)
 #     tailcall_iter(f, iter, iterate(iter, state), f(val, acc))
 # end
@@ -68,7 +66,11 @@ end
 #     tailcall_iter(f, eachindex(x))
 # end
 
-@inline function logdensity_def(d::For{T,F,I}, x::AbstractVector{X}; exec=SequentialEx(simd = true)) where {X,T,F,I<:Tuple{<:AbstractVector}}
+@inline function logdensity_def(
+    d::For{T,F,I},
+    x::AbstractVector{X};
+    exec = SequentialEx(simd = true),
+) where {X,T,F,I<:Tuple{<:AbstractVector}}
     ind = only(d.inds)
     @floop exec for j in eachindex(x)
         local i = getindex(ind, j)
@@ -78,24 +80,27 @@ end
     ℓ
 end
 
-function logdensity_def(d::For, x::AbstractVector; exec=SequentialEx(simd = true)) 
+function logdensity_def(d::For, x::AbstractVector; exec = SequentialEx(simd = true))
     @floop exec for j in eachindex(x)
         local i = (getindex(ind, j) for ind in d.inds)
         local Δℓ = @inbounds logdensity_def(d.f(i...), x[j])
         @reduce ℓ += Δℓ
     end
- 
+
     ℓ
 end
-   
 
-function logdensity_def(d::For{T,F,I}, x::AbstractArray{X}; exec=SequentialEx(simd = true)) where {T,F,I,X}
+function logdensity_def(
+    d::For{T,F,I},
+    x::AbstractArray{X};
[…]
