+++ src/realized.jl
@@ -161,10 +161,18 @@ Base.length(r::RealizedSamples) = length(r.parent)
 
 Base.IteratorSize(r::RealizedSamples) = Base.IteratorSize(r.parent)
 
-function Base.rand(rng::AbstractRNG, ::Type{T}, d::ProductMeasure{G}) where {T, G<:Base.Generator}
+function Base.rand(
+    rng::AbstractRNG,
+    ::Type{T},
+    d::ProductMeasure{G},
+) where {T,G<:Base.Generator}
     RealizedSamples(ResettableRNG(rng), marginals(d))
 end
 
-function Base.rand(rng::AbstractRNG, ::Type, d::For{T,F,I}) where {N,T,F,I<:NTuple{N,<:Base.Generator}}
+function Base.rand(
+    rng::AbstractRNG,
+    ::Type,
+    d::For{T,F,I},
+) where {N,T,F,I<:NTuple{N,<:Base.Generator}}
     RealizedSamples(ResettableRNG(rng), marginals(d))
 end
