+++ src/parameterized/lkj-cholesky.jl
@@ -79,7 +79,6 @@ end
 
 TV.as(d::LKJCholesky) = CorrCholesky(d.k)
 
-
 @inline function basemeasure(d::LKJCholesky{(:k, :η)})
     t = TV.as(d)
     base = Pushforward(t, Lebesgue(ℝ)^TV.dimension(t), False())
