+++ src/distproxy.jl
@@ -1,14 +1,12 @@
 export proxy
 function proxy end
 
-PROXIES = Dict(
-    :Distributions => [
-        :mean
-        :std
-        :entropy
-        :cdf
-    ],
-)
+PROXIES = Dict(:Distributions => [
+    :mean
+    :std
+    :entropy
+    :cdf
+])
 
 for m in keys(PROXIES)
     for f in PROXIES[m]
