+++ src/MeasureTheory.jl
@@ -49,7 +49,16 @@ export ifelse
 
 import MeasureBase: insupport, instance, marginals
 import MeasureBase:
-    testvalue, logdensity_def, density_def, basemeasure, kleisli, params, paramnames, ∫, 𝒹, ∫exp
+    testvalue,
+    logdensity_def,
+    density_def,
+    basemeasure,
+    kleisli,
+    params,
+    paramnames,
+    ∫,
+    𝒹,
+    ∫exp
 import MeasureBase: ≪
 using MeasureBase: BoundedInts, BoundedReals, CountingMeasure, IntegerDomain, IntegerNumbers
 using MeasureBase: weightedmeasure, restrict
