+++ test/runtests.jl
@@ -11,7 +11,7 @@ using MeasureTheory
 using MeasureBase.Interface
 
 using Aqua
-Aqua.test_all(MeasureTheory; ambiguities=false, unbound_args=false)
+Aqua.test_all(MeasureTheory; ambiguities = false, unbound_args = false)
 
 function draw2(μ)
     x = rand(μ)
@@ -19,7 +19,7 @@ function draw2(μ)
     while x == y
         y = rand(μ)
     end
-    return (x,y)
+    return (x, y)
 end
 
 function test_measure(μ)
@@ -28,25 +28,29 @@ end
 
 test_measures = Any[
     # Chain(x -> Normal(μ=x), Normal(μ=0.0))
-    For(3) do j Normal(σ=j) end
-    For(2,3) do i,j Normal(i,j) end
-    Normal() ^ 3
-    Normal() ^ (2,3)
+    For(3) do j
+        Normal(σ = j)
+    end
+    For(2, 3) do i, j
+        Normal(i, j)
+    end
+    Normal()^3
+    Normal()^(2, 3)
     3 * Normal()
     Bernoulli(0.2)
-    Beta(2,3)
-    Binomial(10,0.3)
+    Beta(2, 3)
+    Binomial(10, 0.3)
     Cauchy()
     Dirichlet(ones(3))
     Exponential()
     Gumbel()
     Laplace()
-    LKJCholesky(3,2.0)
-    Multinomial(n=10,p=[0.2,0.3,0.5])
-    NegativeBinomial(5,0.2)
-    Normal(2,3)
+    LKJCholesky(3, 2.0)
+    Multinomial(n = 10, p = [0.2, 0.3, 0.5])
+    NegativeBinomial(5, 0.2)
+    Normal(2, 3)
     Poisson(3.1)
-    StudentT(ν=2.1)    
+    StudentT(ν = 2.1)
     Uniform()
     Counting(Float64)
     Dirac(0.0) + Normal()
@@ -70,9 +74,9 @@ testbroken_measures = Any[
         @info "testing $μ"
         @test_broken test_measure(μ)
     end
-    
+
     @testset "testvalue(::Chain)" begin
-        mc =  Chain(x -> Normal(μ=x), Normal(μ=0.0))
+        mc = Chain(x -> Normal(μ = x), Normal(μ = 0.0))
         r = testvalue(mc)
         @test logdensity_def(mc, Iterators.take(r, 10)) isa AbstractFloat
     end
@@ -81,59 +85,61 @@ end
 @testset "Parameterized Measures" begin
     @testset "Binomial" begin
         D = Binomial{(:n, :p)}
-        par = merge((n=20,),transform(asparams(D, (n=20,)), randn(1)))
+        par = merge((n = 20,), transform(asparams(D, (n = 20,)), randn(1)))
         d = D(par)
-        (n,p) = (par.n, par.p)
+        (n, p) = (par.n, par.p)
         logitp = logit(p)
         probitp = norminvcdf(p)
         y = rand(d)
 
-        ℓ = logdensity_def(Binomial(;n, p), y)
-        @test ℓ ≈ logdensity_def(Binomial(;n, logitp), y)
-        @test ℓ ≈ logdensity_def(Binomial(;n, probitp), y)
+        ℓ = logdensity_def(Binomial(; n, p), y)
+        @test ℓ ≈ logdensity_def(Binomial(; n, logitp), y)
+        @test ℓ ≈ logdensity_def(Binomial(; n, probitp), y)
 
-        @test_broken logdensity_def(Binomial(n,p), CountingMeasure(ℤ[0:n]), x) ≈ binomlogpdf(n,p,x)
[…]
