+++ src/parameterized/poisson.jl
@@ -30,6 +30,6 @@ Base.rand(rng::AbstractRNG, T::Type, d::Poisson{(:λ,)}) = rand(rng, Dists.Poiss
 Base.rand(rng::AbstractRNG, T::Type, d::Poisson{(:logλ,)}) =
     rand(rng, Dists.Poisson(exp(d.logλ)))
 
-@inline function insupport(::Poisson, x) 
+@inline function insupport(::Poisson, x)
     isinteger(x) && x ≥ 0
 end
