+++ src/combinators/affine.jl
@@ -240,10 +240,12 @@ end
 
 # We can't do this until we know we're working with Lebesgue measure, since for
 # example it wouldn't make sense to apply a log-Jacobian to a point measure
-@inline basemeasure(d::Affine{N,L}) where {N,L<:Lebesgue} =
+@inline function basemeasure(d::Affine{N,L}) where {N,L<:Lebesgue}
     weightedmeasure(-logjac(d), d.parent)
-@inline basemeasure(d::Affine{N,L}) where {N,L<:LebesgueMeasure} =
+end
+@inline function basemeasure(d::Affine{N,L}) where {N,L<:LebesgueMeasure}
     weightedmeasure(-logjac(d), d.parent)
+end
 
 logjac(d::Affine) = logjac(getfield(d, :f))
 
