+++ src/parameterized/inverse-gamma.jl
@@ -12,8 +12,9 @@ export InverseGamma
     return xlogy(α + 1, xinv) - xinv - loggamma(α)
 end
 
-Base.rand(rng::AbstractRNG, T::Type, μ::InverseGamma{(:shape,)}) =
+function Base.rand(rng::AbstractRNG, T::Type, μ::InverseGamma{(:shape,)})
     rand(rng, Dists.InverseGamma(μ.shape))
+end
 
 ≪(::InverseGamma, ::Lebesgue{X}) where {X<:Real} = true
 

[…]
