### 2022-05-10 d0962df
CompatHelper: bump compat for MeasureBase to 0.8, (keep existing compat)
