+++ paper/paper.tex
@@ -495,42 +495,42 @@ Finally, kernels are key to expressing \href{https://en.wikipedia.org/wiki/Likel
 
 % \moritz{Ok, I just figured out: A kernel can represent a "random measure" (then it looks like $\kappa(\omega, A)$) or a parametrized distribution, then it looks $\kappa(x, A)$ and then it's the image measure of a parametrized random variable $X_x$. Both cases differ on which of the two arguments of $\kappa$ is related to the $\Omega$ or `rng`. Sorry}
 
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-\section{Markov chains}
+% %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+% \section{Markov chains}
 
-Some kernels $\kappa$ map elements of a state space $X$ to measures on the same space.
+% Some kernels $\kappa$ map elements of a state space $X$ to measures on the same space.
 
-In this case, repeated application of the kernel ``walks'' through $X$, yielding the law of a \href{https://en.wikipedia.org/wiki/Markov_chain}{\emph{Markov chain}}. We support this in \verb|MeasureTheory| through the \verb|Chain| combinator. This takes a kernel (or function) and a starting measure, and can be used with Julia's convenient \verb|do| notation. For example, we can write a Gaussian random walk as
+% In this case, repeated application of the kernel ``walks'' through $X$, yielding the law of a \href{https://en.wikipedia.org/wiki/Markov_chain}{\emph{Markov chain}}. We support this in \verb|MeasureTheory| through the \verb|Chain| combinator. This takes a kernel (or function) and a starting measure, and can be used with Julia's convenient \verb|do| notation. For example, we can write a Gaussian random walk as
 
-\begin{verbatim}
-    mc = Chain(Normal(μ=0.0)) do x Normal(μ=x) end
-\end{verbatim}
+% \begin{verbatim}
+%     mc = Chain(Normal(μ=0.0)) do x Normal(μ=x) end
+% \end{verbatim}
 
-A random sample from a \verb|Chain| is an infinite sequence expressed as a iterator, e.g.
+% A random sample from a \verb|Chain| is an infinite sequence expressed as a iterator, e.g.
 
-\begin{verbatim}
-    julia> r = rand(mc);
+% \begin{verbatim}
+%     julia> r = rand(mc);
 
-    julia> x = Iterators.take(r, 3) |> collect
-    3-element Vector{Any}:
-     -0.4931543737034523
-     -0.5661895116186417
-     -1.3286977670590228
-\end{verbatim}
+%     julia> x = Iterators.take(r, 3) |> collect
+%     3-element Vector{Any}:
+%      -0.4931543737034523
+%      -0.5661895116186417
+%      -1.3286977670590228
+% \end{verbatim}
 
 
 
-Thus, for example, we can evaluate \verb|mc| on the \verb|x| we just computed:
-\begin{verbatim}
-    julia> logdensity(mc, x)
-    -0.4149771036439342
-\end{verbatim}
+% Thus, for example, we can evaluate \verb|mc| on the \verb|x| we just computed:
+% \begin{verbatim}
+%     julia> logdensity(mc, x)
+%     -0.4149771036439342
+% \end{verbatim}
 
-Crucially, the call to \verb|rand| builds a struct containing a random seed, to ensure that repeated realizations of \verb|r| will be identical.
[…]
