### 2022-05-12 37cc997
Steal from abstract
