### 2022-05-12 984b03b
Guides for newcomers
