+++ paper/paper.tex
@@ -560,6 +560,8 @@ Finally, kernels are key to expressing \href{https://en.wikipedia.org/wiki/Likel
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 \section{Densities and integrals}
 
+\chad{I think our adoption of DensityInterface complicates this section - needs some review}
+
 In working with measures, two computations that come up quite often are (Radon-Nikodym) \emph{differentiation} and (Lebesgue) \emph{integration}. It's often important to be able to do things like
 \begin{enumerate}
     \item For measures $\mu$ and $\nu$, obtain $f=\frac{\dd\mu}{\dd\nu}$
