+++ paper/paper.tex
@@ -164,10 +164,10 @@ Of course, this special case is particularly useful. The point is not to disrega
 
 Our work is novel in several ways. \verb|MeasureTheory.jl| has
 \begin{enumerate}
-    \item Explicitly represented \emph{base measures} with the same sophistication as the rest of the system, in particular more than just ``discrete or continuous'';
-    \item A \emph{local} approach for determining \href{https://en.wikipedia.org/wiki/Absolute_continuity}{\emph{absolute continuity}}, which is usually a global characteristic;
-    \item Multiple parameterizations for a given measure, without a proliferation of constructors; and
-    \item Normalization and support constraints held separately from the data-dependent computation, allowing for greater efficiency.
+    \item[$\bullet\!$]Explicitly represented \emph{base measures} with the same sophistication as the rest of the system, in particular more than just ``discrete or continuous'';
+    \item[$\bullet\!$]A \emph{local} approach for determining \href{https://en.wikipedia.org/wiki/Absolute_continuity}{\emph{absolute continuity}}, which is usually a global characteristic;
+    \item[$\bullet\!$]Multiple parameterizations for a given measure, without a proliferation of constructors; and
+    \item[$\bullet\!$]Normalization and support constraints held separately from the data-dependent computation, allowing for greater efficiency.
 \end{enumerate}
 
 Some parts of our approach can been seen in existing systems, though these are still far from universal:
@@ -509,7 +509,7 @@ or
     κ = kernel(Normal; μ=identity, σ=sqrt)
 \end{verbatim}
 
-The latter formulation decomposes the kernel into separate ``parameter maps'', making relationships like the one between \verb|x| and \verb|mean(κ(x))| more explicit.
+The latter formulation decomposes the kernel into separate ``parameter maps'', making relationships between \verb|x| and the parameters of \verb|κ(x)| more explicit.
 
 This could of course also be expressed as a parameterized measure. The difference is that kernels are lighter weight to build and have all the dynamism of functions, while the static nature of parameterized measures can make it easier to express some optimizations.
 
@@ -559,43 +559,43 @@ Finally, kernels are key to expressing \href{https://en.wikipedia.org/wiki/Likel
 % The appropriate $\sigma$-algebra for infinite sequences requires some care. A Markov chain on $X$ has $\sigma$-algebra generated by sets of \emph{sequences with finitely many constraints}.\footnote{That is,  the smallest $\sigma$-algebra containing all sets of sequences which can be written as $\{x \colon x_i \in A_i \mid i \in I\}$ using collections of measurable sets $\{A_i \mid i\in I\}$ indexed by finite index sets $I$, called \href{https://en.wikipedia.org/wiki/Cylinder_set_measure}{\emph{cylinder sets.}}}
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-\section{Densities and integrals}
+% \section{Densities and integrals}
 
-\chad{I think our adoption of DensityInterface complicates this section - needs some review}
+% \chad{I think our adoption of DensityInterface complicates this section - needs some review}
 
[…]
