+++ paper/paper.tex
@@ -520,9 +520,6 @@ This could of course also be expressed as a parameterized measure. The differenc
 % \moritz{Add example where composition of linear Gaussian kernels gives linear Gaussian kernels.}
 % \chad{Let's talk about composition in the product section, otherwise the easy stuff comes after the hard stuff. Update: I added some discussion there, but nothing yet about Gaussian kernels.}
 
-Finally, kernels are key to expressing \href{https://en.wikipedia.org/wiki/Likelihood_function}{\emph{likelihoods}}, which we can use to formulate Bayesian conjugacy properties. We'll discuss this further in Section~\ref{Likelihoods}.
-
-
 % \moritz{Ok, I just figured out: A kernel can represent a "random measure" (then it looks like $\kappa(\omega, A)$) or a parametrized distribution, then it looks $\kappa(x, A)$ and then it's the image measure of a parametrized random variable $X_x$. Both cases differ on which of the two arguments of $\kappa$ is related to the $\Omega$ or `rng`. Sorry}
 
 % %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
