+++ paper/paper.tex
@@ -608,7 +608,8 @@ This could of course also be expressed as a parameterized measure. The differenc
 
 The \href{https://en.wikipedia.org/wiki/Pointwise_product}{\emph{pointwise product}} in probability often describes the fusion of information, in this case the fusion of information from prior and from the observations via the likelihood, but also shows up in related situation, such as \href{https://en.wikipedia.org/wiki/Sensor_fusion}{\emph{sensor fusion}} in signal processing.
 
-Let us momentarily assume the common case of Lebesgue base measure. For a parameter $\theta$ and data $x$, suppose we have a \href{https://en.wikipedia.org/wiki/Prior_probability}{\emph{prior density}} $p(\theta)$ and \href{https://en.wikipedia.org/wiki/Likelihood_function}{\emph{likelihood}} $p(x | \theta)$. Then Bayes's Law gives the posterior density
+To begin, let's consider the case of probability distributions dominated by Lebesgue measure.
+For a parameter $\theta$ and data $x$, suppose we have a \href{https://en.wikipedia.org/wiki/Prior_probability}{\emph{prior density}} $p(\theta)$ and \href{https://en.wikipedia.org/wiki/Likelihood_function}{\emph{likelihood}} $p(x | \theta)$. Then Bayes's Law gives the posterior density
 \[
 p(\theta \mid x) = \frac{p(\theta)\ p(x \mid \theta)}{p(x)}\ .
 \]
@@ -620,9 +621,7 @@ In practice, we rarely know the normalization factor $p(x)$, so we often work in
 {\underbrace{p(\theta)}_\text{prior}}\ 
 {\underbrace{p(x \mid \theta)}_\text{likelihood}} .
 \]
-Because the product does not include normalization, in general it does not integrate to one, and so it's not a probability density function. But there are no such problems as a density for a measure.
-
-
+Because the product does not include normalization, in general it does not integrate to one, and so it's not a probability density function. But there are no such problems as a density for a measure. We now consider this more general formulation.
 
 Given a \emph{parameter space} $\Theta$ and \emph{observation space} $X$, a \verb|Likelihood| $\ell$ consists of a kernel $\ell_κ : \Theta \to \cM(X)$ and an observation $\ell_x \in X$. Given a measure $\rho \in \cM(X)$, $\ell$ can be treated as a function $\Theta \to \RR_{\ge 0}$, given by
 \[
