+++ paper/paper.tex
@@ -383,7 +383,8 @@ $\mu\ll_x \nu$& (finite) & \verb|-Inf|  \\
 \section{Base measures and density recursion}
 It's very useful for users to be able to \emph{call} \verb|logdensity_rel| for arbitrary measures. But \emph{defining} things directly in these terms would require a definition for every pair of possible measures, which is of course intractable.
 
-We'll now discuss the mechanics of defining a measure, followed by {\bf XXX description of the next subsection XXX}
+We'll now discuss the mechanics of defining a measure, followed by 
+\chad{add a description of the next subseciton here}
 
 \subsection*{Defining a measure}
 
