+++ paper/paper.tex
@@ -637,7 +637,7 @@ A likelihood can be treated as a function $\Theta \to \RR_{\ge 0}$, by defining
     \ell(\theta) = \frac{\dd\ell_\kappa(\theta)}{\dd\ell_\beta}(\ell_x)
 \]
 
-Of particular interest for the base measure is the case
+Of particular interest is the case where the base measure $\ell_\beta$ is
 \[
     \ell_\beta = \hat{\theta} \equiv \arg\max_\theta \frac{\dd\ell_\kappa(\theta)}{\dd\rho}(\ell_x)
 \]
