+++ src/combinators/affine.jl
@@ -3,9 +3,9 @@ using LinearAlgebra
 
 const AFFINEPARS = [
     (:μ, :σ)
-    (:μ, :ω)
+    (:μ, :λ)
     (:σ,)
-    (:ω,)
+    (:λ,)
     (:μ,)
 ]
 
@@ -23,17 +23,17 @@ Base.propertynames(d::AffineTransform{N}) where {N} = N
 
 import InverseFunctions: inverse
 
-@inline inverse(f::AffineTransform{(:μ, :σ)}) = AffineTransform((μ = -(f.σ \ f.μ), ω = f.σ))
-@inline inverse(f::AffineTransform{(:μ, :ω)}) = AffineTransform((μ = -f.ω * f.μ, σ = f.ω))
-@inline inverse(f::AffineTransform{(:σ,)}) = AffineTransform((ω = f.σ,))
-@inline inverse(f::AffineTransform{(:ω,)}) = AffineTransform((σ = f.ω,))
+@inline inverse(f::AffineTransform{(:μ, :σ)}) = AffineTransform((μ = -(f.σ \ f.μ), λ = f.σ))
+@inline inverse(f::AffineTransform{(:μ, :λ)}) = AffineTransform((μ = -f.λ * f.μ, σ = f.λ))
+@inline inverse(f::AffineTransform{(:σ,)}) = AffineTransform((λ = f.σ,))
+@inline inverse(f::AffineTransform{(:λ,)}) = AffineTransform((σ = f.λ,))
 @inline inverse(f::AffineTransform{(:μ,)}) = AffineTransform((μ = -f.μ,))
 
 # `size(f) == (m,n)` means `f : ℝⁿ → ℝᵐ`  
 Base.size(f::AffineTransform{(:μ, :σ)}) = size(f.σ)
-Base.size(f::AffineTransform{(:μ, :ω)}) = size(f.ω)
+Base.size(f::AffineTransform{(:μ, :λ)}) = size(f.λ)
 Base.size(f::AffineTransform{(:σ,)}) = size(f.σ)
-Base.size(f::AffineTransform{(:ω,)}) = size(f.ω)
+Base.size(f::AffineTransform{(:λ,)}) = size(f.λ)
 
 function Base.size(f::AffineTransform{(:μ,)})
     (n,) = size(f.μ)
@@ -44,9 +44,9 @@ Base.size(f::AffineTransform, n::Int) = @inbounds size(f)[n]
 
 (f::AffineTransform{(:μ,)})(x) = x + f.μ
 (f::AffineTransform{(:σ,)})(x) = f.σ * x
-(f::AffineTransform{(:ω,)})(x) = f.ω \ x
+(f::AffineTransform{(:λ,)})(x) = f.λ \ x
 (f::AffineTransform{(:μ, :σ)})(x) = f.σ * x + f.μ
-(f::AffineTransform{(:μ, :ω)})(x) = f.ω \ x + f.μ
+(f::AffineTransform{(:μ, :λ)})(x) = f.λ \ x + f.μ
 
 rowsize(x) = ()
 rowsize(x::AbstractArray) = (size(x, 1),)
@@ -78,13 +78,13 @@ end
     return x
 end
 
-@inline function apply!(x, f::AffineTransform{(:ω,),Tuple{F}}, z) where {F<:Factorization}
-    ldiv!(x, f.ω, z)
+@inline function apply!(x, f::AffineTransform{(:λ,),Tuple{F}}, z) where {F<:Factorization}
+    ldiv!(x, f.λ, z)
     return x
 end
 
-@inline function apply!(x, f::AffineTransform{(:ω,)}, z)
-    ldiv!(x, factorize(f.ω), z)
+@inline function apply!(x, f::AffineTransform{(:λ,)}, z)
[…]
