+++ paper/paper.tex
@@ -642,7 +642,7 @@ Of particular interest is the case where the base measure $\ell_\beta$ is
 
 for any non-singular measure $\rho$; this corresponds to the \href{https://en.wikipedia.org/wiki/Likelihood_function#Relative_likelihood_function}{\emph{relative likelihood}}. In \verb|MeasureTheory|, we allow the base measure to be specified, but take as a default the \emph{root measure} of $\ell_\kappa(\theta)$.
 
-Note that a likelihood $\ell$ takes a point in $\Theta$ to an evaluated density in a different space (the observation space $X$). In particular, a likelihood is neither a measure nor a density.
+Note that a likelihood $\ell$ takes a point in $\Theta$ to an evaluated density in a different space (the observation space $X$). In particular, a likelihood is not a measure.
 
 Rather, a given likelihood acts on measures, taking a ``prior'' measure $\mu$ to ``unnormalized posterior'' measure $\mu \odot \ell$. Relative to another measure $\alpha$, this has density
 
