### 2022-05-18 #92 @cscherrer
Hi, :wave:

I'd like to thank @vchuravy and our reviewers @ludgerpaehler and @femtomc for helpful feedback and patience. MeasureTheory has grown and matured steadily since our initial submission, so we had an issue of trying to hit a moving target with a paper that will soon be somewhat frozen. I think we're finally in a position to address this more thoroughly.

We have made some modifications to the repository which should now be visible. We've also made some significant edits to the paper, to make it match the current state of the package. These changes are in a separate branch. I expect @mschauer and I can get these pushed to the `paper` branch very soon, likely within the next day or two. After we merge these, the branch will still be available in case you'd like to see details of exactly what has changed.

I'll update again here when the changes are merged.
