### 2022-05-20 60d763e
Update paper/paper.tex
Co-authored-by: Moritz Schauer <moritzschauer@web.de>
