+++ paper/paper.tex
@@ -139,7 +139,6 @@ We then can work with the ``unnormalized posterior density'' instead, but we wan
 In developments like this using distributions, it's common to begin in terms of a distribution, but then carry around the unnormalized posterior as a function. Fortunately, the meaning of this function is usually clear from context. But structurally, this representation is now divorced from its meaning. It's no longer a Distribution object, so the tools from the original library can no longer help. This is a perspective we will take often.
 
 
-%While it does uniquely identify a probability density, it is not a probability density. 
 
 
 Thus, starting with elements of our class of interest, a simple and common operation has led us to something outside of this class. This is analogous to the way polynomials over the reals lead to the complex numbers. If ``the shortest path between two truths in the real domain passes through the complex domain'' (Hadamard), then the same argument can be made for measures as connection between distributions.
