### 2022-05-20 ece0ff3
add kernel tests
