### 2022-05-20 a194aa1
Update paper/paper.tex
Co-authored-by: Moritz Schauer <moritzschauer@web.de>
