+++ paper/paper.tex
@@ -192,7 +192,7 @@ Likewise, a \emph{measure} assigns each \emph{measurable set} a non-negative qua
 The space $X$ could be the space $\{1,2,3,4,5,6\}$ suitable to model a six-sided die, or $X$ might be the 3-dimensional Euclidean space $\RR^3$ suitable to model ``volume'', to give two examples.
 
 One further such space that deserves mention is traditionally denoted $\Omega$. This is an abstract space connected to the real-world notion of \emph{probability}, which we denote as $\operatorname{Prob}$.
-Computationally, $\Omega$ can be considered to be the set of possible states of a random number generator \verb|rng::AbstractRNG|, which is the source of computational randomness.
+Computationally, $\Omega$ can be considered to be the set of possible initial states of a random number generator \verb|rng::AbstractRNG|, which is the source of computational randomness.
 
 $\Omega$ has a special role of tying random quantities a program produces and their
 distributions together, through the notion of random variables, which are
