### 2022-05-20 9dd7ccf
Update paper/paper.tex
Co-authored-by: Moritz Schauer <moritzschauer@web.de>
