+++ paper/paper.tex
@@ -290,7 +290,8 @@ One says the $\sigma$-algebra is \emph{generated} by these sets.
 Indeed, \verb|Distributions.jl| does not give means to compute probability for arbitrary subsets of floating point numbers,
 but gives a way of computing probability mass just of intervals of the form $(-\infty, x]$ through the function \verb|cdf|.
 
-For spaces that are ``small enough'', a meaningful probability can be assigned to each subset of the space. Unfortunately, this is impossible for any space containing a continuum such as the real line. Just representing a single arbitrary subset of the \verb|Float64|-range, the computational abstraction of the real line, would require staggering 2\,306 Petabytes of memory. This is not practical.
+For spaces that are ``small enough'', a meaningful probability can be assigned to each subset of the space. 
+But even in moderately sized spaces this can become infeasible: Just representing a single arbitrary subset of the \verb|Float64|-range, the computational abstraction of the real line, would require staggering 2\,306 Petabytes of memory. This is not practical. And for any space containing a continuum such as the real line it becomes not impractical but mathematically impossible to assign a meaningful probability to all subsets. 
 
 But even the mathematical measure theory needs to stay clear of the set of all subsets of uncountable spaces, because it turns out that not even volume can be properly defined for all subsets of the Euclidean space.\footnote{Another thing to keep in mind is that, for a function $f\colon X \to Y$ to be measurable requires that ``the inverse image of a measurable set must be measurable''. So allowing more sets to $B\subset Y$ to be measurable requires either allowing more measurable $A\subset X$, or restricting the set of measurable functions.}
 
