### 2022-05-20 3b5dab5
Update paper/paper.tex
Co-authored-by: Moritz Schauer <moritzschauer@web.de>
