### 2022-05-20 4c80cd6
Update paper/paper.tex
Co-authored-by: Moritz Schauer <moritzschauer@web.de>
