+++ paper/paper.tex
@@ -820,7 +820,7 @@ In many cases, we prefer the density (or log-density, really) to be fast to eval
 
 An alternative parameterization of a multivariate Gaussian is in terms of its \href{https://en.wikipedia.org/wiki/Precision_(statistics)}{\emph{precision matrix}}, $\Lambda = \VV[\bX]^{-1}$. Similarly to above, we'll write $\lambda$ for the lower Cholesky factor of $\Lambda$, so $\lambda \lambda^t = \Lambda$. The parameterization is then specified by
 \[
-\bZ = \lambda(\bX - \mu)\ .
+\bZ = \lambda\,(\bX - \mu)\ .
 \]
 
 In this parameterization, $\lambda$ is an \emph{inverse scale} or \emph{rate} parameter.
