+++ paper/paper.tex
@@ -64,6 +64,8 @@
 \newunicodechar{ℓ}{\ensuremath{\ell}}
 \newunicodechar{√}{\ensuremath{\sqrt{}}}
 \newunicodechar{⊗}{\ensuremath{\otimes}}
+\newunicodechar{π}{\ensuremath{\pi}}
+\newunicodechar{ℝ}{\ensuremath{\mathbb{R}}}
 \usepackage{booktabs}
 \usepackage{multirow}
 \renewcommand{\arraystretch}{1.2}
@@ -79,7 +81,7 @@
 \usepackage{inconsolata}
 \usepackage[normalem]{ulem} %\sout
 \usepackage{centernot} %\centernot
-%\usepackage[colorinlistoftodos]{todonotes}
+\usepackage[colorinlistoftodos]{todonotes}
 \usepackage{hyperref}
 \setcounter{page}{1}
 \include{moritzdefs}
@@ -113,7 +115,7 @@ giving a gentle introduction to the concepts in this article.
 
 The task is foremost achieved by recognizing \href{https://en.wikipedia.org/wiki/Measure_(mathematics)}{\emph{measure}} as the central object. This enables us to develop a proper concept of \href{https://en.wikipedia.org/wiki/Probability_density_function}{\emph{densities}} as objects relating measures with each others. As densities provide local perspective on measures, they are the key to efficient implementations.
 
-The need to preserve this computationally so important locality leads to the new notion of \emph{locally-dominated measure} solving the so-called ``base measure problem'' and making work with densities and distributions in Julia easier and more flexible.
+The need to preserve this computationally so important locality leads to the new notion of \emph{locally-dominated measure}, solving the so-called ``base measure problem'' and making work with densities and distributions in Julia easier and more flexible.
 
 %Because it will be new to many who can benefit from this package, this article serves a dual role of a gentle introduction to the concepts of measure theory (the field), as well as an overview of the core capabilities of \verb|MeasureTheory.jl| (the package).
 \end{abstract}
@@ -131,14 +133,12 @@ p(\theta \mid x) = \frac{p(\theta)\ p(x \mid \theta)}{p(x)}\ ,
 the denominator is called the \href{https://en.wikipedia.org/wiki/Marginal_likelihood}{\emph{model evidence}}. Computing it efficiently is often difficult or unfeasible.
 
 Some sampling algorithms circumvent this problem and only require knowing the density up to a constant factor.
-
 We then can work with the ``unnormalized posterior density'' instead, but we want to distinguish it from a proper probability density, because that difference determines which algorithms are available.
 
 
 In developments like this using distributions, it's common to begin in terms of a distribution, but then carry around the unnormalized posterior as a function. Fortunately, the meaning of this function is usually clear from context. But structurally, this representation is now divorced from its meaning. It's no longer a Distribution object, so the tools from the original library can no longer help. This is a perspective we will take often.
 
 
-%While it does uniquely identify a probability density, it is not a probability density. 
 
 
 Thus, starting with elements of our class of interest, a simple and common operation has led us to something outside of this class. This is analogous to the way polynomials over the reals lead to the complex numbers. If ``the shortest path between two truths in the real domain passes through the complex domain'' (Hadamard), then the same argument can be made for measures as connection between distributions.
@@ -146,7 +146,8 @@ Thus, starting with elements of our class of interest, a simple and common opera
 
 
 As a second example, people working in Bayesian modeling sometimes use \href{https://en.wikipedia.org/wiki/Prior_probability#Improper_priors}{\emph{improper priors}}. Whatever one's position on the merits of this, we think it's reasonable to make this approach possible. But an improper prior does not integrate to one, so it's not a distribution.
[…]
