+++ paper/ref.bib
@@ -398,6 +398,14 @@ year = {2009}
     year = 2021
 }
 
+@misc{DensityInterface.jl,
+  author = {Schulz, Oliver and Widmann, David},
+  title = {{DensityInterface.jl}},
+  howpublished = "\url{https://github.com/JuliaMath/DensityInterface.jl}",
+  year = 2021
+}
+
+
 @misc{DynamicIterators.jl,
     author = {Schauer, Moritz},
     title = {{DynamicIterators.jl}},
