### 2022-05-20 #92 @mschauer
@femtomc Do you think #200 addresses your concerns in a way adequate for a small project?
