### 2022-05-20 #92 @cscherrer
Up to you of course, but there are quite a few changes. There's a diff here in case that's helpful:
https://github.com/cscherrer/MeasureTheory.jl/pull/201/files#diff-983b58b8186b0a4ed7f280f258cdab3eb0dd7d5136f8ac361ba982a43cfb7136
