+++ paper/paper.tex
@@ -712,7 +712,7 @@ and
 \]
 
 A very common special case of superposition is a \href{https://en.wikipedia.org/wiki/Spike-and-slab_regression}{\emph{spike and slab prior}}, a mixture of a \href{https://en.wikipedia.org/wiki/Dirac_measure}{\emph{Dirac measure}} (a point mass) with a continuous measure. 
-This is useful for sparse Bayesian modeling, as implemented in the \emph{sparse ZigZag sampler}, described in \cite{arxiv2103.08478} and implemented in \cite{https://doi.org/10.5281/zenodo.3931118} using \verb|MeasureTheory|.
+This is useful for sparse Bayesian modeling, as implemented in the \emph{sparse ZigZag sampler}, described in \cite{bierkens2021sticky} and implemented in \cite{https://doi.org/10.5281/zenodo.3931118} using \verb|MeasureTheory|.
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 \section{Density decomposition}
