+++ paper/ref.bib
@@ -41,13 +41,6 @@ isbn10={0120828200},
   publisher={SIAM}
 }
 
-@misc{arxiv2103.08478,
-Author = {Joris Bierkens and Sebastiano Grazzi and Frank van der Meulen and Moritz Schauer},
-Title = {Sticky {PDMP} samplers for sparse and local inference problems},
-Year = {2021},
-Eprint = {arXiv:2103.08478},
-}
-
 @misc{https://doi.org/10.5281/zenodo.1005091,
   doi = {10.5281/ZENODO.1005091},
   url = {https://zenodo.org/record/1005091},
@@ -213,12 +206,14 @@ ABSTRACT={This review treats the mathematical and algorithmic foundations of non
 
 
 @misc{bierkens2021sticky,
-      title={Sticky {PDMP} samplers for sparse and local inference problems}, 
-      author={Joris Bierkens and Sebastiano Grazzi and Frank van der Meulen and Moritz Schauer},
-      year={2021},
-      eprint={2103.08478},
-      archivePrefix={arXiv},
-      primaryClass={stat.CO}
+  doi = {10.48550/ARXIV.2103.08478},
+  url = {https://arxiv.org/abs/2103.08478},
+  author = {Bierkens,  Joris and Grazzi,  Sebastiano and van der Meulen,  Frank and Schauer,  Moritz},
+  keywords = {Computation (stat.CO),  FOS: Computer and information sciences,  FOS: Computer and information sciences},
+  title = {Sticky PDMP samplers for sparse and local inference problems},
+  publisher = {arXiv},
+  year = {2021},
+  copyright = {arXiv.org perpetual,  non-exclusive license}
 }
 
 @article{Bierkens2018,
