+++ paper/ref.bib
@@ -210,7 +210,7 @@ ABSTRACT={This review treats the mathematical and algorithmic foundations of non
   url = {https://arxiv.org/abs/2103.08478},
   author = {Bierkens,  Joris and Grazzi,  Sebastiano and van der Meulen,  Frank and Schauer,  Moritz},
   keywords = {Computation (stat.CO),  FOS: Computer and information sciences,  FOS: Computer and information sciences},
-  title = {Sticky PDMP samplers for sparse and local inference problems},
+  title = {Sticky {PDMP} samplers for sparse and local inference problems},
   publisher = {arXiv},
   year = {2021},
   copyright = {arXiv.org perpetual,  non-exclusive license}
