### 2022-05-23 #92 @cscherrer
(minor updates to the bibliography)
