### 2022-05-23 b50e55b
better `Pretty.tile` for Affine and AffineTransforms
