+++ src/MeasureTheory.jl
@@ -10,7 +10,6 @@ const TV = TransformVariables
 
 using TransformVariables: asℝ₊, as𝕀, asℝ, transform
 
-
 import Base
 import Distributions
 const Dists = Distributions
