### 2022-05-24 967f6cf
CompatHelper: bump compat for Compat to 4, (keep existing compat)
