+++ Project.toml
@@ -42,7 +42,7 @@ Tricks = "410a4b4d-49e4-4fbc-ab6d-cb71b17b3775"
 [compat]
 Accessors = "0.1"
 ChangesOfVariables = "0.1"
-Compat = "3.35"
+Compat = "3.35, 4"
 ConcreteStructs = "0.2"
 ConstructionBase = "1.3"
 DensityInterface = "0.4"
