### 2022-05-24 533e063
Let's come back to InverseGaussian
