+++ src/parameterized/inverse-gaussian.jl
@@ -4,34 +4,34 @@ export InverseGaussian
 
 @parameterized InverseGaussian()
 
-@kwstruct InverseGaussian()
-@kwstruct InverseGaussian(μ)
-@kwstruct InverseGaussian(μ, λ)
+# @kwstruct InverseGaussian()
+# @kwstruct InverseGaussian(μ)
+# @kwstruct InverseGaussian(μ, λ)
 
-InverseGaussian(μ) = InverseGaussian((μ = μ,))
+# InverseGaussian(μ) = InverseGaussian((μ = μ,))
 
-InverseGaussian(μ, λ) = InverseGaussian((μ = μ, λ = λ))
+# InverseGaussian(μ, λ) = InverseGaussian((μ = μ, λ = λ))
 
-@useproxy InverseGaussian{(:k, :θ)}
+# @useproxy InverseGaussian{(:k, :θ)}
 
-function logdensity_def(d::InverseGaussian{()}, x)
-    return (-3log(x) - (x - 1)^2 / x) / 2
-end
+# function logdensity_def(d::InverseGaussian{()}, x)
+#     return (-3log(x) - (x - 1)^2 / x) / 2
+# end
 
-function logdensity_def(d::InverseGaussian{(:μ,)}, x)
-    μ = d.μ
-    return (-3log(x) - (x - μ)^2 / (μ^2 * x)) / 2
-end
+# function logdensity_def(d::InverseGaussian{(:μ,)}, x)
+#     μ = d.μ
+#     return (-3log(x) - (x - μ)^2 / (μ^2 * x)) / 2
+# end
 
-function logdensity_def(d::InverseGaussian{(:μ, :λ)}, x)
-    μ, λ = d.μ, d.λ
-    return (log(λ) - 3log(x) - λ * (x - μ)^2 / (μ^2 * x)) / 2
-end
+# function logdensity_def(d::InverseGaussian{(:μ, :λ)}, x)
+#     μ, λ = d.μ, d.λ
+#     return (log(λ) - 3log(x) - λ * (x - μ)^2 / (μ^2 * x)) / 2
+# end
 
-function basemeasure(::InverseGaussian)
-    ℓ = static(-log2π / 2)
-    weightedmeasure(ℓ, Lebesgue())
-end
+# function basemeasure(::InverseGaussian)
+#     ℓ = static(-log2π / 2)
+#     weightedmeasure(ℓ, Lebesgue())
+# end
 
 Base.rand(rng::AbstractRNG, T::Type, d::InverseGaussian) = rand(rng, proxy(d))
 
@@ -39,6 +39,7 @@ as(::InverseGaussian) = asℝ₊
 
 insupport(::InverseGaussian, x) = x > 0
 
+# GLM parameterization
 @kwstruct InverseGaussian(μ, ϕ)
 
 function logdensity_def(d::InverseGaussian{(:μ, :ϕ)}, x)
