+++ .github/workflows/ci.yml
@@ -23,6 +23,7 @@ jobs:
       matrix:
         version:
           - '1.6'
+          - '1.7'
         os:
           - ubuntu-latest
           - macOS-latest
