+++ test/runtests.jl
@@ -604,8 +604,8 @@ end
 @testset "IfElseMeasure" begin
     p = rand()
     x = randn()
-    @test logdensityof(MeasureTheory.ifelse(Bernoulli(p), Normal(), Normal()), x) ≈
+    @test logdensityof(MeasureTheory.IfElse.ifelse(Bernoulli(p), Normal(), Normal()), x) ≈
           logdensityof(Normal(), x)
-    @test logdensityof(MeasureTheory.ifelse(Bernoulli(p), Normal(2, 3), Normal()), x) ≈
+    @test logdensityof(MeasureTheory.IfElse.ifelse(Bernoulli(p), Normal(2, 3), Normal()), x) ≈
           logdensityof(p * Normal(2, 3) + (1 - p) * Normal(), x)
 end
