+++ src/combinators/weighted.jl
@@ -1,2 +1,2 @@
 
-TV.as(μ::AbstractWeightedMeasure) = TV.as(μ.base)
+as(μ::AbstractWeightedMeasure) = as(μ.base)

[…]
