+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.16",
   "v0.15",
   "v0.14",
   "v0.13",
@@ -16,5 +17,5 @@ var DOC_VERSIONS = [
   "v0.2",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.15.1";
+var DOCUMENTER_NEWEST = "v0.16.0";
 var DOCUMENTER_STABLE = "stable";
