+++ versions.js
@@ -17,5 +17,5 @@ var DOC_VERSIONS = [
   "v0.2",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.16.0";
+var DOCUMENTER_NEWEST = "v0.16.1";
 var DOCUMENTER_STABLE = "stable";
