### 2022-06-06 01a7c99
CompatHelper: bump compat for Static to 0.7, (keep existing compat)
