+++ Project.toml
@@ -64,7 +64,7 @@ PositiveFactorizations = "0.2"
 PrettyPrinting = "0.3, 0.4"
 Reexport = "1"
 SpecialFunctions = "1, 2"
-Static = "0.5, 0.6"
+Static = "0.5, 0.6, 0.7"
 StaticArrays = "1.3"
 StatsBase = "0.32, 0.33"
 StatsFuns = "0.9, 1"
