### 2022-06-08 e2cd838
insupport(d::OrthoLebesgue, x)
