+++ src/parameterized/mvnormal.jl
@@ -82,35 +82,3 @@ insupport(d::MvNormal, x) = insupport(proxy(d), x)
 @inline proxy(d::MvNormal{(:Λ,),Tuple{C}}) where {C<:Cholesky}    = affine((λ = d.Λ.L,), Normal()^supportdim(d))
 @inline proxy(d::MvNormal{(:μ, :Σ),Tuple{C}}) where {C<:Cholesky} = affine((μ = d.μ, σ = d.Σ.L), Normal()^supportdim(d))
 @inline proxy(d::MvNormal{(:μ, :Λ),Tuple{C}}) where {C<:Cholesky} = affine((μ = d.μ, λ = d.Λ.L), Normal()^supportdim(d))
-
-# # Shouldn't need these, but it was having trouble inferring the type
-# @inline testvalue(d::MvNormal{(:Σ,),Tuple{C}}) where {C<:Cholesky}    = zeros(size(d.Σ, 1))
-# @inline testvalue(d::MvNormal{(:Λ,),Tuple{C}}) where {C<:Cholesky}    = zeros(size(d.Λ, 1))
-# @inline testvalue(d::MvNormal{(:μ, :Σ),Tuple{C}}) where {C<:Cholesky} = zeros(size(d.Σ, 1))
-# @inline testvalue(d::MvNormal{(:μ, :Λ),Tuple{C}}) where {C<:Cholesky} = zeros(size(d.Λ, 1))
-
-
-# function MvNormal(nt::NamedTuple{(:μ,)})
-#     dim = size(nt.μ)
-#     affine(nt, Normal() ^ dim)
-# end
-
-# function MvNormal(nt::NamedTuple{(:σ,)})
-#     dim = colsize(nt.σ)
-#     affine(nt, Normal() ^ dim)
-# end
-
-# function MvNormal(nt::NamedTuple{(:λ,)})
-#     dim = rowsize(nt.λ)
-#     affine(nt, Normal() ^ dim)
-# end
-
-# function MvNormal(nt::NamedTuple{(:μ, :σ,)})
-#     dim = colsize(nt.σ)
-#     affine(nt, Normal() ^ dim)
-# end
-
-# function MvNormal(nt::NamedTuple{(:μ, :λ,)})
-#     dim = rowsize(nt.λ)
-#     affine(nt, Normal() ^ dim)
-# end
