### 2022-06-15 93d9d27
CompatHelper: bump compat for MeasureBase to 0.10, (keep existing compat)
