### 2022-06-18 34e8850
Improve vartransforms docs
