### 2022-06-20 bb2df8d
CompatHelper: bump compat for MeasureBase to 0.11, (keep existing compat)
