### 2022-06-21 bdd5ffa
CompatHelper: bump compat for MeasureBase to 0.12, (keep existing compat)
