### 2022-06-23 55086f8
So many `rand` methods, do we really need all of these?
