+++ test/runtests.jl
@@ -12,7 +12,18 @@ using MeasureBase.Interface
 using MeasureTheory: kernel
 using Aqua
 using IfElse
-Aqua.test_all(MeasureTheory; ambiguities = false, unbound_args = false)
+
+# Aqua._test_ambiguities(
+#     Aqua.aspkgids(MeasureTheory);
+#     exclude = [Random.AbstractRNG],
+#     # packages::Vector{PkgId};
+#     # color::Union{Bool, Nothing} = nothing,
+#     # exclude::AbstractArray = [],
+#     # # Options to be passed to `Test.detect_ambiguities`:
+#     # detect_ambiguities_options...,
+# )
+
+Aqua.test_all(MeasureBase; ambiguities = false)
 
 function draw2(μ)
     x = rand(μ)
@@ -23,8 +34,8 @@ function draw2(μ)
     return (x, y)
 end
 
-x = randn(10,3)
-Σ = cholesky(x'*x)
+x = randn(10, 3)
+Σ = cholesky(x' * x)
 Λ = cholesky(inv(Σ))
 σ = MeasureTheory.getL(Σ)
 λ = MeasureTheory.getL(Λ)
@@ -611,7 +622,7 @@ end
 @testset "IfElseMeasure" begin
     p = rand()
     x = randn()
-    
+
     @test let
         a = logdensityof(IfElse.ifelse(Bernoulli(p), Normal(), Normal()), x)
         b = logdensityof(Normal(), x)
