### 2022-06-30 #92 @cscherrer
Thanks @mschauer . I was worried about violating some DOI uniqueness requirements, but I think your point about referring to a specific version makes sense.
