### 2022-06-30 #92 @mschauer
No worries, there is no uniqueness requirement for DOIs, I also linked them:
