### 2022-06-30 #92 @vchuravy
The JOSS docs are clearer on that:

> We ask that the authors issue a new tagged release of the software (if changed), and archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or other). The authors then post the version number and archive DOI in the REVIEW issue. The handling editor executes the pre-publication steps, and pings the EiCs for final processing.
