### 2022-07-01 #92 @cscherrer
Ok here it is:
[![DOI](https://zenodo.org/badge/233287642.svg)](https://zenodo.org/badge/latestdoi/233287642)
That's for `cscherrer/MeasureTheory.jl: v0.16.4`
