+++ paper/ref.bib
@@ -323,7 +323,7 @@ keywords = {measure kernels, conditional distributions, Probabilistic programs, 
 @article{LEWANDOWSKI20091989,
 abstract = {We extend and improve two existing methods of generating random correlation matrices, the onion method of Ghosh and Henderson [S. Ghosh, S.G. Henderson, Behavior of the norta method for correlated random vector generation as the dimension increases, ACM Transactions on Modeling and Computer Simulation (TOMACS) 13 (3) (2003) 276–294] and the recently proposed method of Joe [H. Joe, Generating random correlation matrices based on partial correlations, Journal of Multivariate Analysis 97 (2006) 2177–2189] based on partial correlations. The latter is based on the so-called D-vine. We extend the methodology to any regular vine and study the relationship between the multiple correlation and partial correlations on a regular vine. We explain the onion method in terms of elliptical distributions and extend it to allow generating random correlation matrices from the same joint distribution as the vine method. The methods are compared in terms of time necessary to generate 5000 random correlation matrices of given dimensions.},
 author = {Lewandowski, Daniel and Kurowicka, Dorota and Joe, Harry},
-doi = {https://doi.org/10.1016/j.jmva.2009.04.008},
+doi = {10.1016/j.jmva.2009.04.008},
 issn = {0047-259X},
 journal = {Journal of Multivariate Analysis},
 keywords = { Correlation matrix, Onion method, Partial correlation,Dependence vines},
