+++ paper/ref.bib
@@ -34,6 +34,7 @@ isbn10={0120828200},
   title={Julia: {A} fresh approach to numerical computing},
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
   journal={SIAM review},
+  doi={10.1137/141000671}
   volume={59},
   number={1},
   pages={65--98},
@@ -169,25 +170,15 @@ URL = {https://doi.org/10.1214/18-AOS1715}
 }
 
 @ARTICLE{10.3389/fphy.2021.663457,
-  
-AUTHOR={Krauth, Werner},   
-	 
+AUTHOR={Krauth, Werner},  
 TITLE={Event-Chain {M}onte {C}arlo: Foundations, Applications, and Prospects},      
-	
 JOURNAL={Frontiers in Physics},      
-	
 VOLUME={9},      
-
 PAGES={229},     
-	
 YEAR={2021},      
-	  
 URL={https://www.frontiersin.org/article/10.3389/fphy.2021.663457},       
-	
 DOI={10.3389/fphy.2021.663457},      
-	
 ISSN={2296-424X},   
-   
 ABSTRACT={This review treats the mathematical and algorithmic foundations of non-reversible Markov chains in the context of event-chain Monte Carlo (ECMC), a continuous-time lifted Markov chain that employs the factorized Metropolis algorithm. It analyzes a number of model applications and then reviews the formulation as well as the performance of ECMC in key models in statistical physics. Finally, the review reports on an ongoing initiative to apply ECMC to the sampling problem in molecular simulation, i.e., to real-world models of peptides, proteins, and polymers in aqueous solution.}
 }
 
@@ -386,6 +377,7 @@ year = {2009}
   title={Stochastic lambda calculus and monads of probability distributions},
   author={Ramsey, Norman and Pfeffer, Avi},
   booktitle={Proceedings of the 29th ACM SIGPLAN-SIGACT symposium on Principles of programming languages},
+  doi={10.1145/503272.503288}
   pages={154--165},
   year={2002}
 }
