+++ paper/ref.bib
@@ -34,7 +34,7 @@ isbn10={0120828200},
   title={Julia: {A} fresh approach to numerical computing},
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
   journal={SIAM review},
-  doi={10.1137/141000671}
+  doi={10.1137/141000671},
   volume={59},
   number={1},
   pages={65--98},
