+++ paper/ref.bib
@@ -377,7 +377,7 @@ year = {2009}
   title={Stochastic lambda calculus and monads of probability distributions},
   author={Ramsey, Norman and Pfeffer, Avi},
   booktitle={Proceedings of the 29th ACM SIGPLAN-SIGACT symposium on Principles of programming languages},
-  doi={10.1145/503272.503288}
+  doi={10.1145/503272.503288},
   pages={154--165},
   year={2002}
 }
