+++ paper/ref.bib
@@ -34,7 +34,7 @@ isbn10={0120828200},
   title={Julia: {A} fresh approach to numerical computing},
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
   journal={SIAM review},
-  doi={10.1137/141000671}
+  doi={10.1137/141000671},
   volume={59},
   number={1},
   pages={65--98},
@@ -377,7 +377,7 @@ year = {2009}
   title={Stochastic lambda calculus and monads of probability distributions},
   author={Ramsey, Norman and Pfeffer, Avi},
   booktitle={Proceedings of the 29th ACM SIGPLAN-SIGACT symposium on Principles of programming languages},
-  doi={10.1145/503272.503288}
+  doi={10.1145/503272.503288},
   pages={154--165},
   year={2002}
 }
